### 2024-09-27 #7284 @kthyng
@adrienbanse
* Please shorten your paper by about half. You might consider removing much of the Methodology section and after that. You can get a feel for the format of the JOSS paper here: https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain
