### 2024-09-30 c7bdfdf
paper: shorten paper
