+++ paper/paper.md
@@ -28,101 +28,37 @@ bibliography: paper.bib
 
 Recent advancements in generative AI (GenAI), particularly Large Language Models (LLMs), have led to significant increases in computational power and energy demands. These models are not only becoming larger and more complex but are also being widely deployed across various platforms, reaching millions of users and developers. This widespread use has escalated the overall energy consumption and environmental impact associated with these technologies.
 
-To address these concerns, we introduce EcoLogits, an open-source Python package designed to evaluate the environmental footprint of using GenAI models. EcoLogits enables users to estimate energy consumption and assess various environmental impacts — including carbon emissions, abiotic resource depletion, and primary energy consumption — associated with each API request to an GenAI service provider. Specifically, the tool implements a bottom-up Life Cycle Assessment (LCA) methodology, outlined in this paper and extensively described in the package documentation, covering both the usage impacts from direct energy consumption and the embodied impacts from hardware production and lifecycle. EcoLogits adheres to high standards of open-source practices, including versioning, automated testing, and comprehensive code documentation.
+To address these concerns, we introduce EcoLogits, an open-source Python package designed to evaluate the environmental footprint of using GenAI models. EcoLogits estimates energy consumption and multi-criteria environmental impacts of GenAI API requests using a bottom-up Life Cycle Assessment methodology, considering both direct energy use and hardware production impacts. EcoLogits adheres to high standards of open-source practices, including versioning, automated testing, and comprehensive code documentation.
 
 # Statement of need
 
-While the recent developments in GenAI and LLMs offer tools of unprecedented quality, there is a general trend of increased energy consumption associated with these technologies. It is a known fact that the training process of large neural networks yields non-negligible carbon emissions [@lacoste_quantifying_2019; @patterson_carbon_2021]. Some industry actors communicate about the emissions of training their own models, such as Meta that reported the carbon emissions of training Llama models [@touvron_llama_2023; @touvron_llama_2023-1; @llama3modelcard_2024].
+While the recent developments in GenAI offer tools of unprecedented quality, there is a general trend of increased energy consumption associated with these technologies [@de_vries_growing_2023]. It is well known that training large neural networks yields non-negligible carbon emissions [@lacoste_quantifying_2019; @patterson_carbon_2021]. Some industry actors communicate about the emissions of training their own models, such as Meta that reported the carbon emissions of Llama models [@touvron_llama_2023; @touvron_llama_2023-1; @llama3modelcard_2024].
 
-The major difference in recent years is that GenAI models are not only large models but are also used daily by millions of people on user-friendly websites such as OpenAI’s ChatGPT or via easy-to-use API clients written in popular languages such as Python or Javascript. The literature agrees on the fact that the share of carbon emissions from the inference is increasing [@patterson_carbon_2021; @desislavov_trends_2023] although these proportions vary among providers, with AWS reporting 80-90%, Google 60% and Meta 33% [@luccioni_power_2024].
+The major difference in recent years is that GenAI models are not only large models but are also widely used by millions of people through wep application such as OpenAI's ChatGPT or via APIs. The literature agrees on the fact that the share of carbon emissions from the inference is increasing [@patterson_carbon_2021; @desislavov_trends_2023] although these proportions vary among providers, with AWS reporting 80-90%, Google 60% and Meta 33% [@luccioni_power_2024].
 
[…]
