+++ paper/paper.bib
@@ -136,24 +136,6 @@
   url = {https://github.com/meta-llama/llama3/blob/main/MODEL_CARD.md}
 }
 
-
-@inproceedings{luccioni_power_2024,
-	title = {Power {Hungry} {Processing}: {Watts} {Driving} the {Cost} of {AI} {Deployment}?},
-	shorttitle = {Power {Hungry} {Processing}},
-	url = {http://arxiv.org/abs/2311.16863},
-	doi = {10.1145/3630106.3658542},
-	abstract = {Recent years have seen a surge in the popularity of commercial AI products based on generative, multi-purpose AI systems promising a unified approach to building machine learning (ML) models into technology. However, this ambition of ``generality'' comes at a steep cost to the environment, given the amount of energy these systems require and the amount of carbon that they emit. In this work, we propose the first systematic comparison of the ongoing inference cost of various categories of ML systems, covering both task-specific (i.e. finetuned models that carry out a single task) and `general-purpose' models, (i.e. those trained for multiple tasks). We measure deployment cost as the amount of energy and carbon required to perform 1,000 inferences on representative benchmark dataset using these models. We find that multi-purpose, generative architectures are orders of magnitude more expensive than task-specific systems for a variety of tasks, even when controlling for the number of model parameters. We conclude with a discussion around the current trend of deploying multi-purpose generative ML systems, and caution that their utility should be more intentionally weighed against increased costs in terms of energy and emissions. All the data from our study can be accessed via an interactive demo to carry out further exploration and analysis.},
-	urldate = {2024-07-17},
-	booktitle = {The 2024 {ACM} {Conference} on {Fairness}, {Accountability}, and {Transparency}},
-	author = {Luccioni, Alexandra Sasha and Jernite, Yacine and Strubell, Emma},
-	month = jun,
-	year = {2024},
-	note = {arXiv:2311.16863 [cs]},
-	keywords = {Inference},
-	pages = {85--99},
-	file = {arXiv Fulltext PDF:/Users/perso/Zotero/storage/B9944C6A/Luccioni et al. - 2024 - Power Hungry Processing Watts Driving the Cost of.pdf:application/pdf;arXiv.org Snapshot:/Users/perso/Zotero/storage/J23YEXW7/2311.html:text/html},
-}
-
 @article{ligozat_unraveling_2022,
 	author = {Ligozat, Anne-Laure and Lefevre, Julien and Bugeau, Aurélie and Combaz, Jacques},
 	title = {Unraveling the Hidden Environmental Impacts of AI Solutions for Environment Life Cycle Assessment of AI Solutions},
@@ -361,27 +343,4 @@
   year         = {1996},
   langid       = {english},
   keywords     = {{ACV} - {LCA}}
-}
-
-@misc{ademe_base_empreinte,
-  title   = {Base Empreinte®},
-  author  = {ADEME},
-  url     = {https://base-empreinte.ademe.fr/},
-  note    = {[Online; accessed 8. Aug. 2024]}
-}
-
-@misc{uptimeinstitute_2024,
-	author = {Jacqueline Davis, John O'Brien},
-	title = {Large data centers are mostly more efficient, analysis confirms},
-	howpublished = {\url{https://journal.uptimeinstitute.com/large-data-centers-are-mostly-more-efficient-analysis-confirms/}},
-	year = {2024},
-	note = {[Accessed 08-08-2024]},
-}
-
-@misc{uptimeinstitute_2022,
-	author = {Jacqueline Davis, Daniel Bizo, Andy Lawrence, Dr. Owen Rogers, Max Smolaks, Lenny Simon, Douglas Donnellan},
[…]
