+++ paper/paper.md
@@ -44,7 +44,7 @@ In parallel to the large deployment of energy-consuming models in the industry, 
 
 The **Attributional Life Cycle Assessment (A-LCA)** methodology defined by the ISO 14044 standard [@klopffer_life_1997; @boustead_lca_1996; @hunt_lca_1996] is used to properly estimate potential environmental impacts of products, processes, projects or services. In EcoLogits, we apply this methodology to estimate the impacts of a request to an GenAI inference service. The methodology encompasses multi-phases and multi-criteria environmental impacts.
 
-In our approach we consider the usage phase accounting for the direct energy consumption [@llm-perf-leaderboard; @optimum-benchmark] and the embodied phase accounting for the hardware production, including raw material extraction, manufacturing and transportation [@boaviztapi]. We do not include the end-of-life phase for lack of data and transparency on e-waste collection and recycling [@bordage_digital_2021; @forti_global_2020]. EcoLogits reports an estimation of the direct energy consumption in addition to other environmental impacts criteria such as *Global Warming Potential* [@ipcc_2013], *Abiotic Resource Depletion* [@van_oers_abiotic_2020] and *Primary Energy* consumption [@frischknecht_cumulative_2015]. Multi-criteria impact assessment helps prevent pollution shifting from single-criterion optimisation. 
+In our approach we consider the usage phase accounting for the direct energy consumption [@llm-perf-leaderboard; @optimum-benchmark] and the embodied phase accounting for the hardware production, including raw material extraction, manufacturing and transportation [@boaviztapi]. We do not include the end-of-life phase for lack of data and transparency on e-waste collection and recycling [@bordage_digital_2021; @forti_global_2020]. EcoLogits reports an estimation of the direct energy consumption in addition to other environmental impacts criteria such as *Global Warming Potential* [@ipcc_2013], *Abiotic Depletion Potential* [@van_oers_abiotic_2020] and *Primary Energy* consumption [@frischknecht_cumulative_2015]. Multi-criteria impact assessment helps prevent pollution shifting from single-criterion optimisation. 
 
 To estimate the impacts of a GenAI service we use a **bottom-up modeling approach**, i.e. we assess the environmental impacts of all components that compose the service individually and then aggregate and allocate these impacts to a request. The scope of our methodology encompasses cloud compute primitives and data center extra equipment such as cooling (see \autoref{fig:boundaries}). We focus our study on high-performance GPU-accelerated cloud instances since they are commonly used for GenAI inference tasks. We do not account for other impacts related to model training, networking transfers or end-user devices.
 
