### 2024-10-01 160e759
paper: minor fixes
