### 2024-10-22 #7284 @kthyng
@samuelrince @adrienbanse Sorry for my delay, I've been traveling and had a deadline to wrap up too. Yes paper looks good though please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. ("u.k." for example but there may be more).
