### 2024-10-28 9f4a59a
test: add ambiguous model name tests
