+++ tests/test_model_repository.py
@@ -1,20 +1,10 @@
-import pytest
-
-from ecologits.model_repository import ModelRepository, Model
+from ecologits.model_repository import ModelRepository
 
 
 def test_create_model_repository_default():
     models = ModelRepository.from_json()
     assert isinstance(models, ModelRepository)
     assert models.find_model(provider="openai", model_name="gpt-3.5-turbo") is not None
-
-
-@pytest.mark.skip
-def test_create_model_repository_from_scratch():
-    models = ModelRepository([
-        Model(provider="provider-test", name="model-test")
-    ])
-    assert models.find_model(provider="provider-test", model_name="model-test") is not None
 
 
 def test_find_unknown_provider():
@@ -25,3 +15,11 @@ def test_find_unknown_provider():
 def test_find_unknown_model_name():
     models = ModelRepository.from_json()
     assert models.find_model(provider="openai", model_name="model-test") is None
+
+
+def test_ambiguous_names():
+    models = ModelRepository.from_json()
+    assert models.find_model(provider="openai", model_name="gpt-4").name == "gpt-4"
+    assert models.find_model(provider="openai", model_name="gpt-4-turbo").name == "gpt-4-turbo"
+    assert models.find_model(provider="openai", model_name="gpt-4o").name == "gpt-4o"
+    assert models.find_model(provider="openai", model_name="gpt-35-turbo").name == "gpt-35-turbo"
