+++ tests/cassettes/test_google/test_google_stream_chat.yaml
@@ -17,60 +17,62 @@ interactions:
       User-Agent:
       - python-requests/2.32.3
       x-goog-api-client:
-      - genai-py/0.7.2 gl-python/3.11.8 grpc/1.64.1 gax/2.19.1
+      - genai-py/0.7.2 gl-python/3.9.18 grpc/1.67.0 gax/2.21.0
       x-goog-request-params:
       - model=models/gemini-1.5-flash
     method: POST
     uri: https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:streamGenerateContent?%24alt=json%3Benum-encoding%3Dint
   response:
     body:
-      string: !!binary |
-        H4sIAAAAAAAC/1SQvQ7CMAyE9zxF5JkBRliZWVC3isFqDI1o46pxpUpV3p2kJf3JEFn3Xc52yklp
-        DRU6Yw0KebjpMipaT/OdGDshJxFkKYod9rJ5lzPt6mgRGtMjKGqCHQlr/TpteT03lMwtG2qyPWQD
-        vK2zvn4SenbRdlmBdYbGqJxVzp5TYfD4oQcJxqVwHR26nttOCv6Su/MwL3VdsnZfcMD/ViAs2BxJ
-        6hlUUD8AAAD//9Lh5aJ5ICrk56QoJCUmZxcAsUJ5YrFCSmlxSaUC0EqF8vyiPB2FktLk7NQUhcTy
-        xEqFzDyFkoxUsHqF/DQwO7GkJDNZhw4RARcpTkxLLakMSizJzEtH9SWyH4GhVpKanl9UiYgKWOgU
-        5SclJmXmZJaA5AzhUrU6BA2yoJZB5tQyyNCAOJOgrFiaJWdznOnZyAyangEAAAD//6JLek7LL0rP
-        LwGaoZCYm5lSXKKQnJ9UnppUDE7RufklGbqpiSDZpJzEvOzUkmI9hZDUPN3K1MQiXWBWGE3HIzwd
-        GxvjTMcmRtB0DAAAAP//7JS7EcJADERbUQEaIj4VEFCGDWJ8g866kXQMdI8sk5KSQAG7ye57X/nx
-        kbkQAj0ai8ZDUrarm5UGrgjmvY4cau5NZii+gZNbShs4Pj6Rps9Nrg4umXfp5wkThSEis2tZNgB7
-        1lEYgcstjB4d9yDEWrmQBju49PzJ+HEy9ruPZBy2bzJeAAAA///s1j0OgCAMhuGrcADjqo7ewwX/
-        gEjEtBDj7f0qzo4uugIpy/skfUUGRxcHi+5D3kZmwkTELzWbNI1FtqPYp5tNr41C0qTEBNuQPMIu
-        rqt9csZGxYk2coze/AEFOCq7tVtb8JFH2ssKNIY9/zhA4CImOWpHP42v02jqh2W+um2cAAAA///s
-        l0EOwjAMBL/iB6BKSJSP8ILQuLSiiSXHlejvWZuKG1cucI2UtQ872c032MDTDk9qqjc3rRabiEuq
-        GJYdk/Ju7972u52TkRfDPXfUwl7r+THMxgWLhB7XFscpIxVsVe7oAhkwEBTOtSEnIql4sEgqITg6
-        x9fCZGskCp1Xdl1F7s5qM04g1YGcopmx7+BZdaBxVd0Cspj2x+vH8Tqezp/56vudrycAAAD//+yX
[…]
