+++ tests/cassettes/test_google/test_google_chat.yaml
@@ -17,40 +17,60 @@ interactions:
       User-Agent:
       - python-requests/2.32.3
       x-goog-api-client:
-      - genai-py/0.7.2 gl-python/3.11.8 grpc/1.64.1 gax/2.19.1
+      - genai-py/0.7.2 gl-python/3.9.18 grpc/1.67.0 gax/2.21.0
       x-goog-request-params:
       - model=models/gemini-1.5-flash
     method: POST
     uri: https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?%24alt=json%3Benum-encoding%3Dint
   response:
     body:
-      string: !!binary |
-        H4sIAAAAAAAC/61Wy67URhDd36/ozIaNMwJFUUKyiAgiUhbZABKKcln02GW7mXaX1d0eXwddiX/I
-        Kr/Hl3Cq2p6ZezfJAiTA0496nDp1qj/eGLOrbWhcYzOl3U/mL6wY81H/lT0OmULGxraExdHGfDlb
-        /ny8+saRTHdyaffK22hNbxtj/WyXZA5EwVjTRLIDxQqfnYvezC73+O4Jpk07eW+4NTMCo+inlA2+
-        ynZz3m1tyDZlV1tvUuboKO3NG67M3MNF6sm0POFaxlcDG4th35iDrY8j/po81UeCUURlXIDpaLoI
-        NwPjQvz86d9kbIb1yjhERD4b746wRHAZlr15C7OerBxGoMnMHEOlzlKOdkymjXahptLQZfkQbUL+
-        cOtxk1m8q0nYF4AUmaNLCeuHxWQ3yFYi3+7Nr1M2iQfKvQudsQfGb1wDirNG7oIgOfcujRTlyBh5
-        cIkEJ9ucUMEp0v423Ia3vc0muK7PlSnF0VJkOZlmlEJu1+w5Jo38yuaJXS0Y/x40nwEJY7lSqGue
-        fBOeIMreIiXZb4nUWBaPCsAGfU8oxCUd2K+d9XvzcoqOk0OlumBnzTRLesWFXAU0SFvJQgAqZNxL
-        5vOnfyR7MDiaA+fsSdFAfCZwpgOwrjYG4T8QSLKVarO3CCsmLWcoflKmcYQjwViwlbBNwxzVhAuZ
-        NZspHAPPQUF9kfTmbL1EmPvIU9frKbSK+DYnh/KHbA5+AhBTRqitX4wXhsNVeADQfq0MMqp7SRkd
-        M5MBycAHI8CIPzeggbC7tk7ueUoSYSneltED4A/UWbhi03meN0gSt+DCQJ4bV5cib1UvvgUE9GVl
-        bnfv8E1SY7QgIurQbXXvfPPL7U5w0LCfpLWN0R3AET3Z05PGBDoJmEJyDtKb+OFANZfRuJVgIrFM
-        TYMDdbQnKb8ckgS4JsTdOu8REULYIxTpv0S2ut0V7LeIrxpuS3ucfNqAokCxW7Rqb9SZX6q1eAA1
-        ixTCMYtGCZoLIrZ1vza29eCmdVG8jaMcFGx6rGiWly44gPrIV8gq5HzIfvQ6FCKkkWMu+XyjtxF/
-        KJ3SQJPobkQdSwMJ5BxJUALY2jOIrYb6ecFurbqnOuseWSTkPTjwTqRQ70/BNE55jdh41kXYdH/z
[…]
