+++ tests/test_google.py
@@ -4,7 +4,7 @@ import google.generativeai as genai
 
 @pytest.mark.vcr
 def test_google_chat(tracer_init):
-    genai.configure(transport='rest')
+    genai.configure(transport="rest")
     model = genai.GenerativeModel("gemini-1.5-flash")
     response = model.generate_content("Write a story about a magic backpack.")
     assert len(response.text) > 0
@@ -16,7 +16,7 @@ def test_google_chat(tracer_init):
 @pytest.mark.vcr
 @pytest.mark.asyncio
 async def test_google_async_chat(tracer_init):
-    genai.configure(transport='rest')
+    genai.configure(transport="rest")
     model = genai.GenerativeModel("gemini-1.5-flash")
     response = await model.generate_content_async("Write a story about a magic backpack.")
     assert len(response.text) > 0
@@ -25,7 +25,7 @@ async def test_google_async_chat(tracer_init):
 
 @pytest.mark.vcr
 def test_google_stream_chat(tracer_init):
-    genai.configure(transport='rest')
+    genai.configure(transport="rest")
     model = genai.GenerativeModel("gemini-1.5-flash")
     stream = model.generate_content(
         "Write a story about a magic backpack.", 
@@ -40,7 +40,7 @@ def test_google_stream_chat(tracer_init):
 @pytest.mark.vcr
 @pytest.mark.asyncio
 async def test_google_async_stream_chat(tracer_init):
-    genai.configure(transport='rest')
+    genai.configure(transport="rest")
     model = genai.GenerativeModel("gemini-1.5-flash")
     stream = await model.generate_content_async(
         "Write a story about a magic backpack.",
