+++ ecologits/tracers/utils.py
@@ -4,8 +4,7 @@ from ecologits.electricity_mix_repository import electricity_mixes
 from ecologits.impacts.llm import compute_llm_impacts
 from ecologits.impacts.modeling import Impacts
 from ecologits.log import logger
-from ecologits.impacts.modeling import Impacts
-from ecologits.model_repository import models, ArchitectureTypes
+from ecologits.model_repository import ArchitectureTypes, models
 
 
 def _avg(value_range: tuple) -> float:
