+++ ecologits/impacts/llm.py
@@ -3,7 +3,7 @@ from math import ceil
 from typing import Any, Optional
 
 from ecologits.impacts.dag import DAG
-from ecologits.impacts.modeling import GWP, PE, ADPe, Embodied, Energy, Impacts, Usage, ValueOrRange
+from ecologits.impacts.modeling import GWP, PE, ADPe, Embodied, Energy, Impacts, Usage
 from ecologits.range_value import RangeValue, ValueOrRange
 
 MODEL_QUANTIZATION_BITS = 4
