### 2024-10-28 #7284 @jaywonchung
Hi @Fei-Tao! I'm not very familiar with JOSS and its review process. Could you point me to a document or provide a quick overview of what is expected of the reviewer, a typical expected time commitment, and a rough due date for the review? I'll be traveling a lot in November, so I need to plan well or in an unfortunate case turn down the review request :(
