### 2024-10-28 #7284 @Fei-Tao
@jaywonchung no problem, please find the review criteria in the following link. https://joss.readthedocs.io/en/latest/review_criteria.html
