+++ ecologits/utils/range_value.py
@@ -63,5 +63,10 @@ class RangeValue(BaseModel):
         else:
             return self.min > other
 
+    def __format__(self,format_spec:str)-> str:
+        return f"{format(self.min,format_spec)} to {format(self.max,format_spec)}"
+
+
+
 
 ValueOrRange = Union[int, float, RangeValue]
