### 2024-11-06 cae919a
fix: range can now be included in f-strings
