+++ tests/test_modeling_impacts.py
@@ -3,8 +3,8 @@ import operator
 import pytest
 
 from ecologits.impacts.modeling import Impact, Energy, GWP, ADPe, PE
-from ecologits.exceptions import ModelingError
-from ecologits.range_value import RangeValue
+from ecologits.utils.exceptions import ModelingError
+from ecologits.utils.range_value import RangeValue
 
 
 impact_config = dict(
