### 2024-11-06 b3913c3
fix: happy ruff happy life
