### 2024-11-07 #7284 @Fei-Tao
Hi @vict0rsch, Thanks for asking. JOSS does not enforce hard constraint. It is recommended to complete the review in a month.
