### 2024-11-12 #7471 @editorialbot
[REVIEW]: EcoLogits: Evaluate the Environmental Impacts of Generative AI
**Submitting author:** @adrienbanse (<a href="http://orcid.org/0000-0002-4456-6618">Adrien Banse</a>)
**Repository:** https://github.com/genai-impact/ecologits
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.1
