+++ ecologits/tracers/utils.py
@@ -3,8 +3,8 @@ from typing import Optional
 from ecologits.electricity_mix_repository import electricity_mixes
 from ecologits.impacts.llm import compute_llm_impacts
 from ecologits.impacts.modeling import Impacts
+from ecologits.log import logger
 from ecologits.model_repository import ArchitectureTypes, models
-from ecologits.utils.log import logger
 
 
 def _avg(value_range: tuple) -> float:
