+++ ecologits/impacts/modeling.py
@@ -2,7 +2,7 @@ from functools import total_ordering
 
 from pydantic import BaseModel
 
-from ecologits.utils.exceptions import ModelingError
+from ecologits.exceptions import ModelingError
 from ecologits.utils.range_value import ValueOrRange
 
 
