### 2024-11-21 #7471 @jaywonchung
https://github.com/genai-impact/ecologits/issues/95
