### 2024-11-22 f2347d1
chore: setup mypy
