### 2024-11-22 039595d
chore: setup mypy
