+++ pyproject.toml
@@ -73,6 +73,7 @@ pytest-dotenv = "^0.5.2"
 pytest-asyncio = "^0.23.6"
 numpy = "^2.0.0"
 toml = "^0.10.2"
+mypy = "^1.13.0"
 
 
 [tool.poetry.group.docs]
@@ -92,6 +93,11 @@ mike = "^2.1.2"
 [pytest]
 env_files = [".env"]
 
+[tool.mypy]
+exclude = [
+    "ecologits/impacts/*",
+    "ecologits/tracers/*"
+]
 
 [tool.ruff]
 select = [
