### 2024-11-22 cd6f4c5
refactor: check models is not none to init repository
