### 2024-11-22 06a899b
refactor: make electricity mix default to WOR
