### 2024-11-22 d0f28e5
refactor: remove incompatible type assignment for mistral
