+++ ecologits/_ecologits.py
@@ -1,7 +1,7 @@
 import importlib.metadata
 import importlib.util
 from dataclasses import dataclass, field
-from typing import Optional, Union
+from typing import Optional, Union, Type
 
 from packaging.version import Version
 
@@ -27,12 +27,17 @@ def init_anthropic_instrumentor() -> None:
 
 def init_mistralai_instrumentor() -> None:
     if importlib.util.find_spec("mistralai") is not None:
+        from ecologits.tracers.mistralai_tracer_v0 import MistralAIInstrumentor as MistralAIInstrumentorV0
+        from ecologits.tracers.mistralai_tracer_v1 import MistralAIInstrumentor as MistralAIInstrumentorV1
+
+        MistralAIInstrumentor: Union[Type[MistralAIInstrumentorV0], Type[MistralAIInstrumentorV1]]
+
         version = Version(importlib.metadata.version("mistralai"))
         if version < Version("1.0.0"):
             logger.warning("MistralAI client v0.*.* will soon no longer be supported by EcoLogits.")
-            from ecologits.tracers.mistralai_tracer_v0 import MistralAIInstrumentor
+            MistralAIInstrumentor = MistralAIInstrumentorV0
         else:
-            from ecologits.tracers.mistralai_tracer_v1 import MistralAIInstrumentor
+            MistralAIInstrumentor = MistralAIInstrumentorV1
 
         instrumentor = MistralAIInstrumentor()
         instrumentor.instrument()
