### 2024-11-22 4f6de6e
refactor: typing of class attributes for DAG
