+++ pyproject.toml
@@ -95,7 +95,8 @@ env_files = [".env"]
 
 [tool.mypy]
 exclude = [
-    "ecologits/impacts/*",
+    "ecologits/impacts/llm.py",
+    "ecologits/impacts/modeling.py",
     "ecologits/tracers/*"
 ]
 
