### 2024-11-22 edb9c4b
refactor: implement __radd__ for RangeValue
