+++ pyproject.toml
@@ -96,7 +96,6 @@ env_files = [".env"]
 [tool.mypy]
 exclude = [
     "ecologits/impacts/llm.py",
-    "ecologits/impacts/modeling.py",
     "ecologits/tracers/*"
 ]
 
