### 2024-11-22 0a2fe85
refactor: change comparison argument for impact model
