+++ ecologits/impacts/modeling.py
@@ -34,21 +34,21 @@ class Impact(BaseModel):
             unit=self.unit
         )
 
-    def __eq__(self, other: "Impact") -> bool:
+    def __eq__(self, other: object) -> bool:
         if not isinstance(other, Impact):
             raise ModelingError(f"Error occurred, cannot compare an Impact with {type(other)}.")
         if self.type != other.type:
             raise ModelingError(f"Error occurred, cannot compare a {self.type} Impact with {other.type} Impact.")
         return self.value == other.value
 
-    def __le__(self, other: "Impact") -> bool:
+    def __le__(self, other: object) -> bool:
         if not isinstance(other, Impact):
             raise ModelingError(f"Error occurred, cannot compare an Impact with {type(other)}.")
         if self.type != other.type:
             raise ModelingError(f"Error occurred, cannot compare a {self.type} Impact with {other.type} Impact.")
         return self.value <= other.value
 
-    def __ge__(self, other: "Impact") -> bool:
+    def __ge__(self, other: object) -> bool:
         if not isinstance(other, Impact):
             raise ModelingError(f"Error occurred, cannot compare an Impact with {type(other)}.")
         if self.type != other.type:
