### 2024-11-22 12da5a6
refactor: typing of internal results when computing llm impacts
