### 2024-11-22 73a4eb2
refactor: accurate typing of add method for impacts
