### 2024-11-22 a9ee4de
refactor: fix typing for model parameters in dag function
