### 2024-11-22 991f02d
refactor: cannot use RangeValue as a min value
