### 2024-11-22 2e0e003
refactor: prefer using official typing for Steam and AsyncStream
