### 2024-11-22 481fbdc
refactor: remove try-catch on import error.
