+++ ecologits/tracers/openai_tracer.py
@@ -1,25 +1,15 @@
 import time
 from typing import Any, Callable, Union
 
-from pydantic import BaseModel
+from openai import AsyncStream, Stream
+from openai.resources.chat import AsyncCompletions, Completions
+from openai.types.chat import ChatCompletion as _ChatCompletion
+from openai.types.chat import ChatCompletionChunk as _ChatCompletionChunk
 from wrapt import wrap_function_wrapper  # type: ignore[import-untyped]
 
 from ecologits._ecologits import EcoLogits
 from ecologits.impacts import Impacts
 from ecologits.tracers.utils import llm_impacts
-
-try:
-    from openai import AsyncStream, Stream
-    from openai.resources.chat import AsyncCompletions, Completions
-    from openai.types.chat import ChatCompletion as _ChatCompletion
-    from openai.types.chat import ChatCompletionChunk as _ChatCompletionChunk
-except ImportError:
-    AsyncStream = object()
-    Stream = object()
-    AsyncCompletions = object()
-    Completions = object()
-    _ChatCompletion = BaseModel
-    _ChatCompletionChunk = BaseModel
 
 
 PROVIDER = "openai"
