### 2024-11-22 db2f461
refactor: remove optional on electricity zone
