### 2024-11-22 2caf422
refactor: use check on parameters instead of arch type
