+++ ecologits/tracers/utils.py
@@ -1,10 +1,11 @@
-from typing import Optional
+from typing import Optional, Union
 
 from ecologits.electricity_mix_repository import electricity_mixes
 from ecologits.impacts.llm import compute_llm_impacts
 from ecologits.impacts.modeling import Impacts
 from ecologits.log import logger
-from ecologits.model_repository import ArchitectureTypes, models
+from ecologits.model_repository import ArchitectureTypes, models, ParametersMoE
+from ecologits.utils.range_value import ValueOrRange
 
 
 def _avg(value_range: tuple) -> float:
@@ -37,7 +38,7 @@ def llm_impacts(
         logger.debug(f"Could not find model `{model_name}` for {provider} provider.")
         return None
 
-    if model.architecture.type == ArchitectureTypes.MOE:
+    if isinstance(model.architecture.parameters, ParametersMoE):
         model_total_params = model.architecture.parameters.total
         model_active_params = model.architecture.parameters.active
     else:
