### 2024-11-23 9e753c6
refactor: cast type and add exception when computing multiple dags
