+++ ecologits/impacts/llm.py
@@ -12,14 +12,13 @@ GPU_ENERGY_ALPHA = 8.91e-8
 GPU_ENERGY_BETA = 1.43e-6
 GPU_LATENCY_ALPHA = 8.02e-4
 GPU_LATENCY_BETA = 2.23e-2
-GPU_MEMORY = 80     # GB
+GPU_MEMORY = 80  # GB
 GPU_EMBODIED_IMPACT_GWP = 143
 GPU_EMBODIED_IMPACT_ADPE = 5.1e-3
 GPU_EMBODIED_IMPACT_PE = 1828
 
-
 SERVER_GPUS = 8
-SERVER_POWER = 1     # kW
+SERVER_POWER = 1  # kW
 SERVER_EMBODIED_IMPACT_GWP = 3000
 SERVER_EMBODIED_IMPACT_ADPE = 0.24
 SERVER_EMBODIED_IMPACT_PE = 38000
@@ -28,16 +27,15 @@ HARDWARE_LIFESPAN = 5 * 365 * 24 * 60 * 60
 
 DATACENTER_PUE = 1.2
 
-
 dag = DAG()
 
 
 @dag.asset
 def gpu_energy(
-    model_active_parameter_count: float,
-    output_token_count: float,
-    gpu_energy_alpha: float,
-    gpu_energy_beta: float
+        model_active_parameter_count: float,
+        output_token_count: float,
+        gpu_energy_alpha: float,
+        gpu_energy_beta: float
 ) -> float:
     """
     Compute energy consumption of a single GPU.
@@ -56,11 +54,11 @@ def gpu_energy(
 
 @dag.asset
 def generation_latency(
-    model_active_parameter_count: float,
-    output_token_count: float,
-    gpu_latency_alpha: float,
-    gpu_latency_beta: float,
-    request_latency: float,
+        model_active_parameter_count: float,
+        output_token_count: float,
+        gpu_latency_alpha: float,
+        gpu_latency_beta: float,
+        request_latency: float,
 ) -> float:
     """
     Compute the token generation latency in seconds.
@@ -81,8 +79,8 @@ def generation_latency(
 
 @dag.asset
 def model_required_memory(
-    model_total_parameter_count: float,
-    model_quantization_bits: int,
+        model_total_parameter_count: float,
+        model_quantization_bits: int,
 ) -> float:
     """
     Compute the required memory to load the model on GPU.
@@ -99,8 +97,8 @@ def model_required_memory(
 
 @dag.asset
 def gpu_required_count(
-    model_required_memory: float,
-    gpu_memory: float
+        model_required_memory: float,
+        gpu_memory: float
 ) -> int:
     """
     Compute the number of required GPU to store the model.
@@ -117,10 +115,10 @@ def gpu_required_count(
 
 @dag.asset
 def server_energy(
-    generation_latency: float,
-    server_power: float,
-    server_gpu_count: int,
-    gpu_required_count: int
+        generation_latency: float,
+        server_power: float,
+        server_gpu_count: int,
+        gpu_required_count: int
 ) -> float:
     """
     Compute the energy consumption of the server.
@@ -139,10 +137,10 @@ def server_energy(
 
 @dag.asset
 def request_energy(
-    datacenter_pue: float,
-    server_energy: float,
-    gpu_required_count: int,
-    gpu_energy: float
+        datacenter_pue: float,
+        server_energy: float,
+        gpu_required_count: int,
+        gpu_energy: float
 ) -> float:
     """
     Compute the energy consumption of the request.
@@ -161,8 +159,8 @@ def request_energy(
 
 @dag.asset
 def request_usage_gwp(
-    request_energy: float,
-    if_electricity_mix_gwp: float
+        request_energy: float,
[…]
