+++ ecologits/tracers/utils.py
@@ -1,11 +1,10 @@
-from typing import Optional, Union
+from typing import Optional
 
 from ecologits.electricity_mix_repository import electricity_mixes
 from ecologits.impacts.llm import compute_llm_impacts
 from ecologits.impacts.modeling import Impacts
 from ecologits.log import logger
-from ecologits.model_repository import ArchitectureTypes, models, ParametersMoE
-from ecologits.utils.range_value import ValueOrRange
+from ecologits.model_repository import ParametersMoE, models
 
 
 def _avg(value_range: tuple) -> float:
