### 2024-11-23 bd53bf5
chore: configure mypy to ignore docs and tests
