+++ pyproject.toml
@@ -96,13 +96,21 @@ env_files = [".env"]
 [tool.mypy]
 exclude = [
     "ecologits/tracers/mistralai_tracer_v1.py",
-    "ecologits/tracers/mistralai_tracer_v0.py",
     "ecologits/tracers/google_tracer.py",
     "ecologits/tracers/cohere_tracer.py",
     "ecologits/tracers/anthropic_tracer.py",
     "ecologits/tracers/litellm_tracer.py",
-    "ecologits/tracers/huggingface_tracer.py",
+    "ecologits/tracers/huggingface_tracer.py"
 ]
+
+[[tool.mypy.overrides]]
+module = "docs.*"
+ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = "tests.*"
+ignore_errors = true
+
 
 
 [tool.ruff]
