### 2024-11-23 799cdc7
refactor: rollback loading of Mistral instrumentor and add mypy ignore flag
