+++ pyproject.toml
@@ -97,7 +97,6 @@ env_files = [".env"]
 exclude = [
     "ecologits/tracers/google_tracer.py",
     "ecologits/tracers/cohere_tracer.py",
-    "ecologits/tracers/anthropic_tracer.py",
     "ecologits/tracers/litellm_tracer.py",
     "ecologits/tracers/huggingface_tracer.py"
 ]
