### 2024-11-23 af94f8f
refactor: typing issue with generators
