### 2024-11-23 8c4efc3
refactor: make impacts optional
