+++ pyproject.toml
@@ -95,7 +95,6 @@ env_files = [".env"]
 
 [tool.mypy]
 exclude = [
-    "ecologits/tracers/litellm_tracer.py",
     "ecologits/tracers/huggingface_tracer.py"
 ]
 
