### 2024-11-23 f6a78e6
refactor: typing issue with generators
