### 2024-11-23 f14a894
fix: typo in exception
