### 2024-11-23 ada6250
fix: use object for other type (similar to comparison)
