### 2024-11-26 38ddf5d
feat: wcf impacts computation (missing data)
