+++ ecologits/impacts/llm.py
@@ -3,7 +3,7 @@ from math import ceil
 from typing import Any, Optional
 
 from ecologits.impacts.dag import DAG
-from ecologits.impacts.modeling import GWP, PE, ADPe, Embodied, Energy, Impacts, Usage
+from ecologits.impacts.modeling import GWP, PE, ADPe, WCF, Embodied, Energy, Impacts, Usage
 from ecologits.utils.range_value import RangeValue, ValueOrRange
 
 MODEL_QUANTIZATION_BITS = 4
@@ -214,6 +214,28 @@ def request_usage_pe(
 
 
 @dag.asset
+def request_usage_wcf(
+    request_energy: float,
+    wue_on_site: float,
+    wue_off_site: float,
+    datacenter_pue: float,
+) -> float:
+    """
+    Compute the Water Consumption Footprint (WCF) usage impact of the request.
+
+    Args:
+        request_energy: Energy consumption of the request in kWh.
+        wue_on_site: On-site Water Usage Effectiveness in L/kWh.
+        wue_off_site: Off-site Water Usage Efectiveness in L/kWh (electricity generation).
+        datacenter_pue: PUE of the datacenter.
+
+    Returns:
+        The WCF usage impact of the request in L.
+    """
+    return request_energy * (wue_on_site + datacenter_pue * wue_off_site)
+
+
+@dag.asset
 def server_gpu_embodied_gwp(
     server_embodied_gwp: float,
     server_gpu_count: float,
@@ -339,6 +361,26 @@ def request_embodied_pe(
     return (generation_latency / server_lifetime) * server_gpu_embodied_pe
 
 
+@dag.asset
+def request_embodied_wcf(
+    generation_latency: float,
+    server_lifetime: float,
+    wcf_manufacturing: float
+) -> float:
+    """
+    Compute the Water Consumption Footprint (WCF) embodied impact of the request.
+
+    Args:
+        generation_latency: Token generation latency in seconds.
+        server_lifetime: Lifetime duration of the server in seconds.
+        wcf_manufacturing: Water Consumption Footprint for manufacturing the server in L.
+
+    Returns:
+        The WCF embodied impact of the request in L.
+    """
+    return wcf_manufacturing * generation_latency / server_lifetime
+
+
 def compute_llm_impacts_dag(
     model_active_parameter_count: float,
     model_total_parameter_count: float,
@@ -347,6 +389,9 @@ def compute_llm_impacts_dag(
     if_electricity_mix_adpe: float,
     if_electricity_mix_pe: float,
     if_electricity_mix_gwp: float,
+    wue_on_site: float,
+    wue_off_site: float,
+    wcf_manufacturing: float,
     model_quantization_bits: Optional[int] = MODEL_QUANTIZATION_BITS,
     gpu_energy_alpha: Optional[float] = GPU_ENERGY_ALPHA,
     gpu_energy_beta: Optional[float] = GPU_ENERGY_BETA,
@@ -375,6 +420,9 @@ def compute_llm_impacts_dag(
         if_electricity_mix_adpe: ADPe impact factor of electricity consumption of kgSbeq / kWh (Antimony).
         if_electricity_mix_pe: PE impact factor of electricity consumption in MJ / kWh.
         if_electricity_mix_gwp: GWP impact factor of electricity consumption in kgCO2eq / kWh.
+        wue_on_site: On-site Water Usage Effectiveness in L/kWh.
+        wue_off_site: Off-site Water Usage Efectiveness in L/kWh (electricity generation).
+        wcf_manufacturing: Water Consumption Footprint for manufacturing the server in L.
         model_quantization_bits: Number of bits used to represent the model weights.
[…]
