+++ ecologits/impacts/modeling.py
@@ -128,6 +128,27 @@ class PE(Impact):
     unit: str = "MJ"
 
 
+class WCF(Impact):
+    """
+    Water Consumption Footprint (WCF) impact.
+
+    Info:
+        Amount of water “evaporated, transpired, incorporated into products or crops, 
+        or otherwise removed from the immediate water environment" (source: Macknick 
+        et al., "A review of operational water consumption and withdrawal factors for
+        electricity generating technologies")
+
+    Attributes:
+        type: WCF
+        name: Water Consumption Footprint
+        value: WCF value
+        unit: Liter (L)
+    """
+    type: str = "WCF"
+    name: str = "Water Consumption Footprint"
+    unit: str = "L"
+
+
 class Phase(BaseModel):
     """
     Base impact phase data model.
@@ -154,6 +175,7 @@ class Usage(Phase):
         gwp: Global Warming Potential (GWP) usage impact
         adpe: Abiotic Depletion Potential for Elements (ADPe) usage impact
         pe: Primary Energy (PE) usage impact
+        wcf: Water Consumption Footprint (WCF) usage impact
     """
     type: str = "usage"
     name: str = "Usage"
@@ -161,6 +183,7 @@ class Usage(Phase):
     gwp: GWP
     adpe: ADPe
     pe: PE
+    wcf: WCF
 
 
 class Embodied(Phase):
@@ -176,12 +199,14 @@ class Embodied(Phase):
         gwp: Global Warming Potential (GWP) embodied impact
         adpe: Abiotic Depletion Potential for Elements (ADPe) embodied impact
         pe: Primary Energy (PE) embodied impact
+        wcf: Water Consumption Footprint (WCF) embodied impact
     """
     type: str = "embodied"
     name: str = "Embodied"
     gwp: GWP
     adpe: ADPe
     pe: PE
+    wcf: WCF
 
 
 class Impacts(BaseModel):
@@ -193,6 +218,7 @@ class Impacts(BaseModel):
         gwp: Total Global Warming Potential (GWP) impact
         adpe: Total Abiotic Depletion Potential for Elements (ADPe) impact
         pe: Total Primary Energy (PE) impact
+        wcf: Total Water Consumption Footprint (WCF) impact
         usage: Impacts for the usage phase
         embodied: Impacts for the embodied phase
     """
@@ -200,5 +226,6 @@ class Impacts(BaseModel):
     gwp: GWP
     adpe: ADPe
     pe: PE
+    wcf: WCF
     usage: Usage
     embodied: Embodied
