### 2024-11-26 genai-impact/ecologits#95 @samuelrince
Hello Jae-Won (@jaywonchung),

First, thank you very much for taking the time to review our work!

I wanted to let you know that we have taken your remarks into account, and we are actively working on the various fixes and improvements you suggested.

We expect to get back to you around mid-December with all the updates to the paper, code, and documentation.
