+++ ecologits/impacts/llm.py
@@ -214,25 +214,61 @@ def request_usage_pe(
 
 
 @dag.asset
-def request_usage_wcf(
+def request_server_usage_wcf(
     request_energy: float,
     wue_on_site: float,
+) -> float:
+    """
+    Compute the Water Consumption Footprint (WCF) usage impact of the request 
+    due to the servers water consumption. 
+
+    Args:
+        request_energy: Energy consumption of the request in kWh.
+        wue_on_site: On-site Water Usage Effectiveness in L/kWh.
+
+    Returns:
+        The server WCF usage impact of the request in L.
+    """
+    return request_energy * wue_on_site
+
+
+@dag.asset
+def request_electricity_usage_wcf(
+    request_energy: float,
     wue_off_site: float,
     datacenter_pue: float,
+) -> float:
+    """
+    Compute the Water Consumption Footprint (WCF) usage impact of the request
+    due to the electricity generation. 
+
+    Args:
+        request_energy: Energy consumption of the request in kWh.
+        wue_off_site: Off-site Water Usage Efectiveness in L/kWh (electricity generation).
+        datacenter_pue: PUE of the datacenter.
+
+    Returns:
+        The electricty WCF usage impact of the request in L.
+    """
+    return request_energy * datacenter_pue * wue_off_site
+
+
+@dag.asset
+def request_usage_wcf(
+    request_server_usage_wcf: float,
+    request_electricity_usage_wcf: float,
 ) -> float:
     """
     Compute the Water Consumption Footprint (WCF) usage impact of the request.
 
     Args:
-        request_energy: Energy consumption of the request in kWh.
-        wue_on_site: On-site Water Usage Effectiveness in L/kWh.
-        wue_off_site: Off-site Water Usage Efectiveness in L/kWh (electricity generation).
-        datacenter_pue: PUE of the datacenter.
+        request_server_usage_wcf: The server WCF usage impact of the request in L.
+        request_electricity_usage_wcf: The electricty WCF usage impact of the request in L.
 
     Returns:
-        The WCF usage impact of the request in L.
+        The total WCF usage impact of the request in L.
     """
-    return request_energy * (wue_on_site + datacenter_pue * wue_off_site)
+    return request_server_usage_wcf + request_electricity_usage_wcf
 
 
 @dag.asset
@@ -361,25 +397,6 @@ def request_embodied_pe(
     return (generation_latency / server_lifetime) * server_gpu_embodied_pe
 
 
-@dag.asset
-def request_embodied_wcf(
-    generation_latency: float,
-    server_lifetime: float,
-    wcf_manufacturing: float
-) -> float:
-    """
-    Compute the Water Consumption Footprint (WCF) embodied impact of the request.
-
-    Args:
-        generation_latency: Token generation latency in seconds.
-        server_lifetime: Lifetime duration of the server in seconds.
-        wcf_manufacturing: Water Consumption Footprint for manufacturing the server in L.
-
-    Returns:
-        The WCF embodied impact of the request in L.
-    """
-    return wcf_manufacturing * generation_latency / server_lifetime
-
 
 def compute_llm_impacts_dag(
     model_active_parameter_count: float,
@@ -391,7 +408,6 @@ def compute_llm_impacts_dag(
     if_electricity_mix_gwp: float,
     wue_on_site: float,
     wue_off_site: float,
-    wcf_manufacturing: float,
     model_quantization_bits: Optional[int] = MODEL_QUANTIZATION_BITS,
     gpu_energy_alpha: Optional[float] = GPU_ENERGY_ALPHA,
[…]
