### 2024-11-27 e0abe08
docs: fix range value mention
