+++ ecologits/tracers/utils.py
@@ -1,10 +1,14 @@
 from typing import Optional
 
-from ecologits.electricity_mix_repository import electricity_mixes
+from ecologits.repositories.electricity_mix_repository import electricity_mixes
+from ecologits.repositories.electricity_wue_repository import electricity_wue_list
 from ecologits.impacts.llm import compute_llm_impacts
 from ecologits.impacts.modeling import Impacts
 from ecologits.log import logger
-from ecologits.model_repository import ArchitectureTypes, models
+from ecologits.repositories.model_repository import ArchitectureTypes, models
+
+
+DEFAULT_ZONE = "WOR"
 
 
 def _avg(value_range: tuple) -> float:
@@ -16,7 +20,7 @@ def llm_impacts(
     model_name: str,
     output_token_count: int,
     request_latency: float,
-    electricity_mix_zone: Optional[str] = "WOR",
+    electricity_zone: Optional[str] = DEFAULT_ZONE,
 ) -> Optional[Impacts]:
     """
     High-level function to compute the impacts of an LLM generation request.
@@ -26,7 +30,7 @@ def llm_impacts(
         model_name: Name of the LLM used.
         output_token_count: Number of generated tokens.
         request_latency: Measured request latency in seconds.
-        electricity_mix_zone: ISO 3166-1 alpha-3 code of the electricity mix zone (WOR by default).
+        electricity_zone: ISO 3166-1 alpha-3 code of the electricity mix zone (WOR by default).
 
     Returns:
         The impacts of an LLM generation request.
@@ -44,13 +48,20 @@ def llm_impacts(
         model_total_params = model.architecture.parameters
         model_active_params = model.architecture.parameters
 
-    electricity_mix = electricity_mixes.find_electricity_mix(zone=electricity_mix_zone)
+    electricity_mix = electricity_mixes.find_electricity_mix(zone=electricity_zone)
+    electricity_wue = electricity_wue_list.find_electricity_wue(zone=electricity_zone)
+
+    # TODO: here handle if defined for one but not the other
     if electricity_mix is None:
-        logger.debug(f"Could not find electricity mix `{electricity_mix_zone}` in the ADEME database")
-        return None
+        logger.warning(f"Could not find zone `{electricity_zone}` in the electricity mixes database (ADEME), world average used instead.")
+        electricity_mix = electricity_mixes.find_electricity_mix(zone=DEFAULT_ZONE)
+    if electricity_wue is None: 
+        logger.warning(f"Could not find zone `{electricity_zone}` in the electricty WUE database (WRI), world average used instead.")
+    
     if_electricity_mix_adpe=electricity_mix.adpe
     if_electricity_mix_pe=electricity_mix.pe
     if_electricity_mix_gwp=electricity_mix.gwp
+    wue_off_site=electricity_wue.wue
     return compute_llm_impacts(
         model_active_parameter_count=model_active_params,
         model_total_parameter_count=model_total_params,
@@ -59,4 +70,5 @@ def llm_impacts(
         if_electricity_mix_adpe=if_electricity_mix_adpe,
         if_electricity_mix_pe=if_electricity_mix_pe,
         if_electricity_mix_gwp=if_electricity_mix_gwp,
+        wue_off_site=wue_off_site,
     )
