### 2024-11-27 db455e5
fix: default WUE if not found
