### 2024-11-28 392ac8e
fix: self object
