### 2024-11-28 dcda2b3
fix: tiktoken without rust compiler
