+++ pyproject.toml
@@ -44,7 +44,7 @@ mistralai = { version = ">=0.4.3,<2.0.0", optional = true }
 anthropic = { version = "^0.29.0", optional = true }
 cohere = {version = "^5.2.5", optional = true}
 huggingface-hub = { version = "^0.22.2",  optional = true }
-tiktoken = { version = "^0.7.0", optional = true }
+tiktoken = { version = "^0.8.0", optional = true }
 aiohttp = { version = "^3.9.3", optional = true }
 minijinja = { version = "^1.0.16", optional = true }
 google-generativeai = { version = "^0.7.0", optional = true }
