+++ ecologits/tracers/mistralai_tracer_v0.py
@@ -47,7 +47,7 @@ def mistralai_chat_wrapper(
         model_name=response.model,
         output_token_count=response.usage.completion_tokens,
         request_latency=request_latency,
-        electricity_mix_zone=EcoLogits.config.electricity_mix_zone
+        electricity_zone=EcoLogits.config.electricity_zone
     )
     if impacts is not None:
         return ChatCompletionResponse(**response.model_dump(), impacts=impacts)
@@ -71,7 +71,7 @@ def mistralai_chat_wrapper_stream(
             model_name=model_name,
             output_token_count=token_count,
             request_latency=request_latency,
-            electricity_mix_zone=EcoLogits.config.electricity_mix_zone
+            electricity_zone=EcoLogits.config.electricity_zone
         )
         if impacts is not None:
             yield ChatCompletionStreamResponse(**chunk.model_dump(), impacts=impacts)
@@ -93,7 +93,7 @@ async def mistralai_async_chat_wrapper(
         model_name=response.model,
         output_token_count=response.usage.completion_tokens,
         request_latency=request_latency,
-        electricity_mix_zone=EcoLogits.config.electricity_mix_zone
+        electricity_zone=EcoLogits.config.electricity_zone
     )
     if impacts is not None:
         return ChatCompletionResponse(**response.model_dump(), impacts=impacts)
@@ -120,7 +120,7 @@ async def mistralai_async_chat_wrapper_stream(
             model_name=model_name,
             output_token_count=token_count,
             request_latency=request_latency,
-            electricity_mix_zone=EcoLogits.config.electricity_mix_zone
+            electricity_zone=EcoLogits.config.electricity_zone
         )
         if impacts is not None:
             yield ChatCompletionStreamResponse(**chunk.model_dump(), impacts=impacts)
