### 2024-11-29 c754230
docs: add scope and proprietary models info
