### 2024-11-29 bed9658
docs: add scope and proprietary models info
