### 2024-11-29 7fcef33
docs: add scope and proprietary models info
