+++ mkdocs.yml
@@ -24,6 +24,7 @@ nav:
           - 'OpenAI': tutorial/providers/openai.md
   - 'Methodology':
       - 'Introduction': methodology/index.md
+      - 'Proprietary models': methodology/proprietary_models.md
       - 'LLM Inference': methodology/llm_inference.md
   - 'API Reference': reference/
   - 'Online Calculator': 'https://huggingface.co/spaces/genai-impact/ecologits-calculator'
@@ -114,6 +115,7 @@ markdown_extensions:
   - tables
   - admonition
   - def_list
+  - footnotes
   - pymdownx.tasklist:
       custom_checkbox: true
   - pymdownx.arithmatex:
