### 2024-11-29 1352650
docs: add scope and proprietary models info
