### 2024-11-29 3457df4
docs: update proprietary models intro
