### 2024-11-29 b412439
docs: detail some hypotheses
