### 2024-11-29 7674efa
docs: detail some hypotheses
