### 2024-11-29 07f95b7
docs: detail some hypotheses
