### 2024-11-29 08a3703
Add mypy static type checker (#96)
Co-authored-by: NP4567-dev <23741953+NP4567-dev@users.noreply.github.com>
