+++ ecologits/impacts/modeling.py
@@ -1,13 +1,16 @@
 from functools import total_ordering
+from typing import TypeVar
 
 from pydantic import BaseModel
 
 from ecologits.exceptions import ModelingError
 from ecologits.utils.range_value import ValueOrRange
 
+Impact = TypeVar("Impact", bound="BaseImpact")
+
 
 @total_ordering
-class Impact(BaseModel):
+class BaseImpact(BaseModel):
     """
     Base impact data model.
 
@@ -22,8 +25,8 @@ class Impact(BaseModel):
     value: ValueOrRange
     unit: str
 
-    def __add__(self, other: "Impact") -> "Impact":
-        if not isinstance(other, Impact):
+    def __add__(self: Impact, other: object) -> Impact:
+        if not isinstance(other, BaseImpact):
             raise ModelingError(f"Error occurred, cannot add an Impact with {type(other)}.")
         if self.type != other.type:
             raise ModelingError(f"Error occurred, cannot add a {self.type} Impact with {other.type} Impact.")
@@ -34,29 +37,29 @@ class Impact(BaseModel):
             unit=self.unit
         )
 
-    def __eq__(self, other: "Impact") -> bool:
-        if not isinstance(other, Impact):
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, BaseImpact):
             raise ModelingError(f"Error occurred, cannot compare an Impact with {type(other)}.")
         if self.type != other.type:
             raise ModelingError(f"Error occurred, cannot compare a {self.type} Impact with {other.type} Impact.")
         return self.value == other.value
 
-    def __le__(self, other: "Impact") -> bool:
-        if not isinstance(other, Impact):
+    def __le__(self, other: object) -> bool:
+        if not isinstance(other, BaseImpact):
             raise ModelingError(f"Error occurred, cannot compare an Impact with {type(other)}.")
         if self.type != other.type:
             raise ModelingError(f"Error occurred, cannot compare a {self.type} Impact with {other.type} Impact.")
         return self.value <= other.value
 
-    def __ge__(self, other: "Impact") -> bool:
-        if not isinstance(other, Impact):
+    def __ge__(self, other: object) -> bool:
+        if not isinstance(other, BaseImpact):
             raise ModelingError(f"Error occurred, cannot compare an Impact with {type(other)}.")
         if self.type != other.type:
             raise ModelingError(f"Error occurred, cannot compare a {self.type} Impact with {other.type} Impact.")
         return self.value >= other.value
 
 
-class Energy(Impact):
+class Energy(BaseImpact):
     """
     Energy consumption.
 
@@ -74,7 +77,7 @@ class Energy(Impact):
     unit: str = "kWh"
 
 
-class GWP(Impact):
+class GWP(BaseImpact):
     """
     Global Warming Potential (GWP) impact.
 
@@ -92,7 +95,7 @@ class GWP(Impact):
     unit: str = "kgCO2eq"
 
 
-class ADPe(Impact):
+class ADPe(BaseImpact):
     """
     Abiotic Depletion Potential for Elements (ADPe) impact.
 
@@ -110,7 +113,7 @@ class ADPe(Impact):
     unit: str = "kgSbeq"
 
 
-class PE(Impact):
+class PE(BaseImpact):
     """
     Primary Energy (PE) impact.
 
