+++ ecologits/_ecologits.py
@@ -32,7 +32,7 @@ def init_mistralai_instrumentor() -> None:
             logger.warning("MistralAI client v0.*.* will soon no longer be supported by EcoLogits.")
             from ecologits.tracers.mistralai_tracer_v0 import MistralAIInstrumentor
         else:
-            from ecologits.tracers.mistralai_tracer_v1 import MistralAIInstrumentor
+            from ecologits.tracers.mistralai_tracer_v1 import MistralAIInstrumentor  # type: ignore[assignment]
 
         instrumentor = MistralAIInstrumentor()
         instrumentor.instrument()
@@ -123,7 +123,7 @@ class EcoLogits:
     @staticmethod
     def init(
         providers: Optional[Union[str, list[str]]] = None,
-        electricity_mix_zone: Optional[str] = "WOR",
+        electricity_mix_zone: str = "WOR",
     ) -> None:
         """
         Initialization static method. Will attempt to initialize all providers by default.

[…]
