### 2024-11-29 d917ab5
docs: mention genai impact non-profit on the top
