### 2024-11-29 d4baace
chore: bump version to 0.5.2
