### 2024-12-02 6275b44
feat: add warnings and errors to output impacts object
