### 2024-12-02 2bdb403
feat: add warnings and errors to output impacts object
