### 2024-12-02 861ecb1
feat: add warnings and errors to output impacts object
