### 2024-12-03 acc7278
feat: change naming to alerts + add docs url
