### 2024-12-03 8f219b3
feat: change naming to alerts + add docs url
