+++ docs/tutorial/alerts.md
@@ -1,0 +1,15 @@
+# Warnings and Errors
+
+## Warnings
+
+### `model-arch-not-released`
+
+
+### `model-arch-multimodal`
+
+
+## Errors
+
+### `model-not-registered`
+
+### `zone-not-registerd`
