+++ mkdocs.yml
@@ -14,6 +14,7 @@ nav:
       - 'Introduction': tutorial/index.md
       - 'Environmental Impacts': tutorial/impacts.md
       - 'Supported providers': tutorial/providers.md
+      - 'Warnings and Errors': tutorial/alerts.md
       - 'Providers':
           - 'Anthropic': tutorial/providers/anthropic.md
           - 'Cohere': tutorial/providers/cohere.md
