### 2024-12-03 27fa25b
feat: change naming to alerts + add docs url
