+++ ecologits/model_repository.py
@@ -5,8 +5,8 @@ from typing import Any, Optional, Union
 
 from pydantic import BaseModel
 
+from ecologits.alerts import AlertMessage
 from ecologits.utils.range_value import ValueOrRange
-from ecologits.warnings_and_errors import BaseWarning
 
 
 class Providers(Enum):
@@ -43,7 +43,7 @@ class Model(BaseModel):
     provider: Providers
     name: str
     architecture: Architecture
-    warnings: list[BaseWarning] = []
+    warnings: list[AlertMessage] = []
     sources: list[str] = []
 
     @property
@@ -54,7 +54,7 @@ class Model(BaseModel):
     def from_json(cls, data: dict[str, Any]) -> "Model":
         warnings = []
         if data["warnings"] is not None:
-            warnings = [BaseWarning.from_code(code) for code in data["warnings"]]
+            warnings = [AlertMessage.from_code(code) for code in data["warnings"]]
         return cls(
             provider=Providers(data["provider"]),
             name=data["name"],
