### 2024-12-05 cdb882f
docs: add more documentation of warnings and errors
