### 2024-12-05 0bb7acf
docs: add more documentation of warnings and errors
