### 2024-12-06 c47215f
chore: update google-generativeai
