+++ tests/cassettes/test_google/test_google_chat.yaml
@@ -1,87 +1,0 @@
-interactions:
-- request:
-    body: "{\n  \"contents\": [\n    {\n      \"parts\": [\n        {\n          \"text\":
-      \"Write a story about a magic backpack.\"\n        }\n      ],\n      \"role\":
-      \"user\"\n    }\n  ],\n  \"generationConfig\": {}\n}"
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '186'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - python-requests/2.32.3
-      x-goog-api-client:
-      - genai-py/0.7.2 gl-python/3.9.18 grpc/1.67.0 gax/2.21.0
-      x-goog-request-params:
-      - model=models/gemini-1.5-flash
-    method: POST
-    uri: https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?%24alt=json%3Benum-encoding%3Dint
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAC/2VXTW/cNhC9+1ewe0kLaA27jd24PSVFDwH6hSZoD00PlDiSmKVIlaR2sw4C9D/0
-        H/aX9M2Q2rXdADFskRrOvPfmDfXxQqlNp72xRmdKm2/UH3ii1Ef5yWvBZ/IZC+sjPJx1zOe95d/H
-        B79jS6YP/NLme6ejVged/LOsXAg76wfVh6i02SPuEulSvRmJdzxeVoY6bFCt7nYz/jdqjtRT1K07
-        quBJ5VFnhbw5cBdSVlOI8tSrkaI6EO2wUzsXDtp3OEa9xTm9I60mHXeUGxzSjTpk26l0nOYxeETu
-        cWLMFDkR4IITUqejIaNyJJ2QcGokWz7EaZw7hpmj/0QpO2zTU/ADHnejnWd+jXS3zEmCTSGPWwLS
-        XnHUPceyWSVU8u/f/yAfH7yh1EU7I4D2e5xzBqDXnMYQCa8fbB7VIUSvUFDmXFKOek5I5HVWPbks
-        D/xAAOGg46RyADgAKCzdePnOv/OvfY52WBCyPSKLpBa/pEU72Y7onDDqnaNNAANhgCRIsB2pz3u7
-        J5Ws2/PBo3UOO9JnXzSq8N2GZRgzggrqqI5ANrY0OINrYXA5Z0+IuESwkKhBIBK4A6IykCpNoK9Z
-        K9wiKlJqoRKVl26HLdYnawD+a2Q/64ESaIcIeiSEVcGIHA0aOqqgGsraOiFXJbCMV0B5msk5cAGq
-        xqKj1t7ruOqJcWOqUc0r8pzN6RGnIJFt8CwogUpqsdNUVNRXgqCcTEagf7MjvNEB63YB0Uu0YUkr
-        eMIQC4flWt5lVn3KqMMdG1lYVaHowwyiyODIg4h9xRbNY7kQMjGEqUQAWA1nDE0arruLemKN/rWI
-        5quyUcUc9UFQcggOqc5UBTpZbzW3LXcpCiogo5HiEfm5bQeyRIukJ8HDh8MW7/MpE/gD+l6JLpGH
[…]
