+++ tests/cassettes/test_google/test_google_stream_chat.yaml
@@ -1,97 +1,0 @@
-interactions:
-- request:
-    body: "{\n  \"contents\": [\n    {\n      \"parts\": [\n        {\n          \"text\":
-      \"Write a story about a magic backpack.\"\n        }\n      ],\n      \"role\":
-      \"user\"\n    }\n  ],\n  \"generationConfig\": {}\n}"
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '186'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - python-requests/2.32.3
-      x-goog-api-client:
-      - genai-py/0.7.2 gl-python/3.9.18 grpc/1.67.0 gax/2.21.0
-      x-goog-request-params:
-      - model=models/gemini-1.5-flash
-    method: POST
-    uri: https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:streamGenerateContent?%24alt=json%3Benum-encoding%3Dint
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAC/1SOMQvCMBCF9/6KcLOVKjjoKo5uxaU4HPaswTRXkisIpf/dJMW2ZjiO9768e9WQ
-        KQUPtLWuUcjDSVVBUWpIM3pshawE4ycFsUMnCzu9YbUHROgTP8HFwMoY5/2+WeIcG4psyzXN+ITG
-        mVDoPTZ0JcFQFOc60DluOyn5TfbMfSp6nKJBWND8O8EYU1o6dSPnNdt4uqFWW53vtof8adC/8qLY
-        QzZmXwAAAP//1NCxDYAwDATAHokdvqMhrENtyCMiWQmyA1G2ZwoKboSbx+HzIjFBE89TRS+3ofYr
-        7aKojfowdIqFonHBeiYlThoOS8zRUTanO+Ovj18AAAD//6JHGCvklwGDrSC/QKG4JLGoWAFolUJx
-        fnImMJhzU1MyE3UUXHNA8ZBSlJoIFFDIT1MozUutKMjJLwLyskrz0nNSIbpyK0sywNGTDFRaUloE
-        dO1QDnwAAAAA///UkrEOgzAMRH/FYkb8RjcmRhZLpMXCTaLENOLvOecjkDrH8Tvf3RPmT7TsUilz
-        dTUjnH2nYmcETa/Ruw+/QUAURk1sJyb4ips2iR9ahzmtAzqfvr38OZQc7GTVq8faUjkQEgRBZHUY
-        G0l1jon2L6/CsfMxx40B1cA/X47Xv87vBgAA///UlEEKwzAMBL+iB4Q+oO/IMRdFyK2IvTGyA3l+
-        ZR/6hEIPuiwIweyuflIeDuZyZWMnQxCzzhB9hh7sMWH2gXlnOWrMQq1ozoNvYkMfRiUSA7icmDWq
-        Bn1s2LC+9bs3o3DBtbAfvMezCyeTeev0yuPkQnqL1hBPj+utRAj+2r4PAAAA///UldENwyAQQ1dh
[…]
