+++ pyproject.toml
@@ -47,7 +47,7 @@ huggingface-hub = { version = "^0.22.2",  optional = true }
 tiktoken = { version = "^0.8.0", optional = true }
 aiohttp = { version = "^3.9.3", optional = true }
 minijinja = { version = "^1.0.16", optional = true }
-google-generativeai = { version = "^0.7.0", optional = true }
+google-generativeai = { version = "^0.8.3", optional = true }
 litellm = {version="^1.41.23", optional = true}
 rapidfuzz = {version = "^3.9.6", optional = true}
 
