+++ tests/cassettes/test_google/test_google_stream_chat.yaml
@@ -1,0 +1,156 @@
+interactions:
+- request:
+    body: "{\n  \"contents\": [\n    {\n      \"parts\": [\n        {\n          \"text\":
+      \"Write a story about a magic backpack.\"\n        }\n      ],\n      \"role\":
+      \"user\"\n    }\n  ],\n  \"generationConfig\": {}\n}"
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '186'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - python-requests/2.32.3
+      x-goog-api-client:
+      - genai-py/0.7.2 gl-python/3.13.0 grpc/1.68.0 gax/2.23.0
+      x-goog-request-params:
+      - model=models/gemini-1.5-flash
+    method: POST
+    uri: https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:streamGenerateContent?%24alt=json%3Benum-encoding%3Dint
+  response:
+    body:
+      string: "[{\n  \"candidates\": [\n    {\n      \"content\": {\n        \"parts\":
+        [\n          {\n            \"text\": \"El\"\n          }\n        ],\n        \"role\":
+        \"model\"\n      }\n    }\n  ],\n  \"usageMetadata\": {\n    \"promptTokenCount\":
+        9,\n    \"totalTokenCount\": 9\n  },\n  \"modelVersion\": \"gemini-1.5-flash\"\n}\n,\r\n{\n
+        \ \"candidates\": [\n    {\n      \"content\": {\n        \"parts\": [\n          {\n
+        \           \"text\": \"ara wasn't your typical twelve-year-old. While others
+        obsessed over\"\n          }\n        ],\n        \"role\": \"model\"\n      }\n
+        \   }\n  ],\n  \"usageMetadata\": {\n    \"promptTokenCount\": 9,\n    \"totalTokenCount\":
+        9\n  },\n  \"modelVersion\": \"gemini-1.5-flash\"\n}\n,\r\n{\n  \"candidates\":
+        [\n    {\n      \"content\": {\n        \"parts\": [\n          {\n            \"text\":
+        \" pop stars and ponies, Elara dreamt of unexplored jungles and hidden cities.
+        Her\"\n          }\n        ],\n        \"role\": \"model\"\n      }\n    }\n
+        \ ],\n  \"usageMetadata\": {\n    \"promptTokenCount\": 9,\n    \"totalTokenCount\":
+        9\n  },\n  \"modelVersion\": \"gemini-1.5-flash\"\n}\n,\r\n{\n  \"candidates\":
+        [\n    {\n      \"content\": {\n        \"parts\": [\n          {\n            \"text\":
+        \" dreams, however, were constantly hampered by her family's crippling poverty.
+        That is, until her grandmother, Nana Elara, passed away, leaving her a\"\n
+        \         }\n        ],\n        \"role\": \"model\"\n      }\n    }\n  ],\n
+        \ \"usageMetadata\": {\n    \"promptTokenCount\": 9,\n    \"totalTokenCount\":
+        9\n  },\n  \"modelVersion\": \"gemini-1.5-flash\"\n}\n,\r\n{\n  \"candidates\":
+        [\n    {\n      \"content\": {\n        \"parts\": [\n          {\n            \"text\":
+        \" worn, leather backpack.\\n\\nIt wasn't particularly stylish.  The leather
+        was scuffed, the stitching frayed in places.  But when Elara touched\"\n          }\n
[…]
