+++ tests/cassettes/test_google/test_google_chat.yaml
@@ -1,0 +1,103 @@
+interactions:
+- request:
+    body: "{\n  \"contents\": [\n    {\n      \"parts\": [\n        {\n          \"text\":
+      \"Write a story about a magic backpack.\"\n        }\n      ],\n      \"role\":
+      \"user\"\n    }\n  ],\n  \"generationConfig\": {}\n}"
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '186'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - python-requests/2.32.3
+      x-goog-api-client:
+      - genai-py/0.7.2 gl-python/3.13.0 grpc/1.68.0 gax/2.23.0
+      x-goog-request-params:
+      - model=models/gemini-1.5-flash
+    method: POST
+    uri: https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?%24alt=json%3Benum-encoding%3Dint
+  response:
+    body:
+      string: "{\n  \"candidates\": [\n    {\n      \"content\": {\n        \"parts\":
+        [\n          {\n            \"text\": \"Elara wasn't your typical twelve-year-old.
+        \ While other kids obsessed over pop stars and social media, Elara dreamt
+        of forgotten languages and unexplored jungles.  This fascination led her,
+        one rainy Tuesday, to a dusty antique shop tucked away on a cobbled side street.\\n\\nThere,
+        amidst chipped teacups and moth-eaten shawls, sat a backpack.  It wasn't particularly
+        beautiful; a drab, olive-green canvas, worn at the seams. But it hummed with
+        a faint, almost imperceptible energy that drew Elara in like a moth to a flickering
+        flame.\\n\\nThe old shopkeeper, a wizened woman with eyes like polished amber,
+        barely glanced up.  \\\"Fifty shillings,\\\" she croaked, her voice like dry
+        leaves rustling.\\n\\nFifty shillings was Elara's entire savings, earned from
+        a year of meticulous babysitting.  She handed it over, her heart hammering
+        a frantic rhythm against her ribs.\\n\\nThe moment Elara's fingers brushed
+        the rough canvas, a jolt, like static electricity, shot up her arm.  Back
+        home, she unzipped the backpack. Inside, nestled amongst the expected compartments,
+        was a small, leather-bound book. Its pages were filled with a language she
+        didn't recognize, yet somehow understood.  It was a catalogue, it seemed,
+        of the backpack's contents.\\n\\nThat night, Elara learned the backpack's
+        magic: it held a seemingly infinite supply of anything she needed, provided
+        she wrote a precise description in the book.  Need a map of the Amazon rainforest?
+        \ A detailed one appeared in the largest compartment.  Craving a perfectly
+        ripe mango? One materialized, glistening and fragrant.  A translation device
+        for ancient Sumerian?  It was there, humming subtly beside a miniature compass
+        that always pointed north, regardless of its orientation.\\n\\nElara's life
+        transformed.  She explored hidden waterfalls in the Scottish Highlands, climbed
+        the slopes of Kilimanjaro (with the backpack thoughtfully providing oxygen
+        and crampons), and even deciphered a centuries-old inscription on a forgotten
+        Mayan temple, using a translation device that magically appeared beside a
+        protective helmet and sturdy boots.\\n\\nBut the magic came with a price.
+        \ Each item requested, each journey undertaken, drained a small, unseen energy
+        from the backpack. The olive-green canvas grew paler, the seams more frayed.
[…]
