+++ ecologits/impacts/llm.py
@@ -496,6 +496,7 @@ def compute_llm_impacts(
     if_electricity_mix_adpe: float,
     if_electricity_mix_pe: float,
     if_electricity_mix_gwp: float,
+    wue_off_site: float,
     request_latency: Optional[float] = None,
     **kwargs: Any
 ) -> Impacts:
@@ -533,8 +534,7 @@ def compute_llm_impacts(
             total_params = [model_total_parameter_count, model_total_parameter_count]
 
     results = {}
-    fields = ["request_energy", "request_usage_gwp", "request_usage_adpe", "request_usage_pe",
-              "request_embodied_gwp", "request_embodied_adpe", "request_embodied_pe"]
+    fields = ["request_energy", "request_usage_gwp", "request_usage_adpe", "request_usage_pe", "request_usage_wcf", "request_embodied_gwp", "request_embodied_adpe", "request_embodied_pe"]
     for act_param, tot_param in zip(active_params, total_params):
         res = compute_llm_impacts_dag(
             model_active_parameter_count=act_param,
