### 2024-12-07 31c76f1
refactor: change alerts to status messages and split impacts and output
