+++ ecologits/tracers/utils.py
@@ -1,15 +1,55 @@
 from typing import Optional
 
-from ecologits.alerts import ModelNotRegisteredError, ZoneNotRegisteredError
+from pydantic import BaseModel
+
 from ecologits.electricity_mix_repository import electricity_mixes
 from ecologits.impacts.llm import compute_llm_impacts
-from ecologits.impacts.modeling import Impacts
+from ecologits.impacts.modeling import GWP, PE, ADPe, Embodied, Energy, Usage
 from ecologits.log import logger
 from ecologits.model_repository import ParametersMoE, models
+from ecologits.status_messages import ErrorMessage, ModelNotRegisteredError, WarningMessage, ZoneNotRegisteredError
 
 
-def _avg(value_range: tuple) -> float:
-    return sum(value_range) / len(value_range)
+class ImpactsOutput(BaseModel):
+    """
+    Impacts output data model.
+
+    Attributes:
+        energy: Total energy consumption
+        gwp: Total Global Warming Potential (GWP) impact
+        adpe: Total Abiotic Depletion Potential for Elements (ADPe) impact
+        pe: Total Primary Energy (PE) impact
+        usage: Impacts for the usage phase
+        embodied: Impacts for the embodied phase
+        warnings: List of warnings
+        errors: List of errors
+    """
+    energy: Optional[Energy] = None
+    gwp: Optional[GWP] = None
+    adpe: Optional[ADPe] = None
+    pe: Optional[PE] = None
+    usage: Optional[Usage] = None
+    embodied: Optional[Embodied] = None
+    warnings: Optional[list[WarningMessage]] = None
+    errors: Optional[list[ErrorMessage]] = None
+
+    @property
+    def has_warnings(self) -> bool:
+        return isinstance(self.warnings, list) and len(self.warnings) > 0
+
+    @property
+    def has_errors(self) -> bool:
+        return isinstance(self.errors, list) and len(self.errors) > 0
+
+    def add_warning(self, warning: WarningMessage) -> None:
+        if self.warnings is None:
+            self.warnings = []
+        self.warnings.append(warning)
+
+    def add_errors(self, error: ErrorMessage) -> None:
+        if self.errors is None:
+            self.errors = []
+        self.errors.append(error)
 
 
 def llm_impacts(
@@ -18,7 +58,7 @@ def llm_impacts(
     output_token_count: int,
     request_latency: float,
     electricity_mix_zone: str = "WOR",
-) -> Optional[Impacts]:
+) -> ImpactsOutput:
     """
     High-level function to compute the impacts of an LLM generation request.
 
@@ -36,7 +76,7 @@ def llm_impacts(
     model = models.find_model(provider=provider, model_name=model_name)
     if model is None:
         logger.debug(f"Could not find model `{model_name}` for {provider} provider.")
-        return Impacts(errors=[ModelNotRegisteredError()])
+        return ImpactsOutput(errors=[ModelNotRegisteredError()])
 
     if isinstance(model.architecture.parameters, ParametersMoE):
         model_total_params = model.architecture.parameters.total
@@ -48,7 +88,7 @@ def llm_impacts(
     electricity_mix = electricity_mixes.find_electricity_mix(zone=electricity_mix_zone)
     if electricity_mix is None:
         logger.debug(f"Could not find electricity mix `{electricity_mix_zone}` in the ADEME database")
-        return Impacts(errors=[ZoneNotRegisteredError()])
+        return ImpactsOutput(errors=[ZoneNotRegisteredError()])
     if_electricity_mix_adpe=electricity_mix.adpe
     if_electricity_mix_pe=electricity_mix.pe
[…]
