### 2024-12-07 59452e2
refactor: change alerts to status messages and split impacts and output
