### 2024-12-07 5cc3607
refactor: change alerts to status messages and split impacts and output
