+++ ecologits/tracers/mistralai_tracer_v0.py
@@ -13,18 +13,17 @@ from mistralai.models.chat_completion import (  # type: ignore[import-not-found]
 from wrapt import wrap_function_wrapper  # type: ignore[import-untyped]
 
 from ecologits._ecologits import EcoLogits
-from ecologits.impacts import Impacts
-from ecologits.tracers.utils import llm_impacts
+from ecologits.tracers.utils import ImpactsOutput, llm_impacts
 
 PROVIDER = "mistralai"
 
 
 class ChatCompletionResponse(_ChatCompletionResponse):
-    impacts: Impacts
+    impacts: ImpactsOutput
 
 
 class ChatCompletionStreamResponse(_ChatCompletionStreamResponse):
-    impacts: Impacts
+    impacts: ImpactsOutput
 
 
 
