### 2024-12-07 86f8be9
docs: update references for impacts object
