### 2024-12-07 ae845aa
style: remove newline in status messages
