### 2024-12-07 f978299
docs: use gpt-4o-mini in all examples
