+++ README.md
@@ -11,7 +11,9 @@
 [![Python version](https://img.shields.io/pypi/pyversions/ecologits)](https://pypi.org/project/ecologits/)
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/drive/1VxrpJ5xuWZKQLsSN12kdqUqkppHRct3G?usp=sharing)
 
-EcoLogits was created and is actively maintained by the **[GenAI Impact](https://genai-impact.org/) non-profit**. Read the full **documentation on [ecologits.ai](https://ecologits.ai/)**.
+EcoLogits was created and is actively maintained by the **[GenAI Impact](https://genai-impact.org/) non-profit**. 
+
+Read the full **documentation on [ecologits.ai](https://ecologits.ai/)**.
 
 
 ## ⚙️ Installation
@@ -34,7 +36,7 @@ EcoLogits.init()
 client = OpenAI(api_key="<OPENAI_API_KEY>")
 
 response = client.chat.completions.create(
-    model="gpt-3.5-turbo",
+    model="gpt-4o-mini",
     messages=[
         {"role": "user", "content": "Tell me a funny joke!"}
     ]
