### 2024-12-07 f0e02bc
docs: use gpt-4o-mini in all examples
