+++ docs/index.md
@@ -73,7 +73,7 @@ EcoLogits.init()
 client = OpenAI(api_key="<OPENAI_API_KEY>")
 
 response = client.chat.completions.create(
-    model="gpt-3.5-turbo",
+    model="gpt-4o-mini",
     messages=[
         {"role": "user", "content": "Tell me a funny joke!"}
     ]
