### 2024-12-07 ee0938b
docs: use gpt-4o-mini in all examples
