+++ ecologits/_ecologits.py
@@ -101,7 +101,7 @@ class EcoLogits:
 
         client = OpenAI(api_key="<OPENAI_API_KEY>")
         response = client.chat.completions.create(
-            model="gpt-3.5-turbo",
+            model="gpt-4o-mini",
             messages=[
                 {"role": "user", "content": "Tell me a funny joke!"}
             ]
