### 2024-12-09 4ad263c
feat: add logger methods to log a message only once
