### 2024-12-09 a1e4827
feat: add logger methods to log a message only once
