### 2024-12-12 569e940
test: update tests with wcf + mypy
