### 2024-12-13 9312d1c
fix: change error type for unknown zone
