### 2024-12-13 616beca
fix: add types to custom logger
