### 2024-12-13 fcd0c05
fix: use "ecologits" as logger name
