### 2024-12-13 636ac8b
tests: add logging and logging once tests
