### 2024-12-13 54f7bc1
ci: update pre-commit workflow
