### 2024-12-16 a9e554a
docs: rework homepage
