### 2024-12-18 1872def
docs: rework environmental impacts page
