### 2024-12-18 cd8f34f
docs: rework environmental impacts page
