+++ docs/tutorial/impacts.md
@@ -1,59 +1,86 @@
 # Environmental Impacts
 
-Environmental impacts are reported for each request in the [`ImpactsOutput`][tracers.utils.ImpactsOutput] pydantic model and features multiple [criteria](#criteria) such as the [energy](#energy) and [global warming potential](#global-warming-potential-gwp) per phase ([usage](#usage) or [embodied](#embodied)) as well as the total impacts.
+Environmental impacts are reported for each request in the **[`ImpactsOutput`][tracers.utils.ImpactsOutput]** that features multiple [impact criteria](#impact-criteria) such as [energyc onsumption](#energy) or the [global warming potential](#global-warming-potential-gwp) per phase ([usage](#usage) or [embodied](#embodied)) as well as the total impacts. It also contains potential [warnings and errors](warnings_and_errors.md) that can occur during the calculation.
 
-To learn more on how we estimate the environmental impacts and what are our hypotheses go to the [methodology](../methodology/index.md) section.
+!!! note "To learn more on how we estimate the environmental impacts and what are our hypotheses go to the [methodology](../methodology/index.md) section."
 
-```python title="Structure of Impacts model"
+
+## Impacts Output
+
+The [`ImpactsOutput`][tracers.utils.ImpactsOutput] is structured the following way:
+
+```python
 from ecologits.tracers.utils import ImpactsOutput
 from ecologits.impacts.modeling import ADPe, Embodied, Energy, GWP, PE, Usage
 
+
 ImpactsOutput(
-    energy=Energy(), # (1)!
-    gwp=GWP(),
-    adpe=ADPe(),
-    pe=PE(),
-    usage=Usage( # (2)!
+    energy=Energy(),    # Total energy consumed
+    gwp=GWP(),          # Total global warming potential (or GHG emissions)
+    adpe=ADPe(),        # Total abiotic resource depletion
+    pe=PE(),            # Total energy consumed from primary sources
+    usage=Usage( # (1)!
         energy=Energy(),
         gwp=GWP(),
         adpe=ADPe(),
         pe=PE(),
     ),
-    embodied=Embodied( # (3)!
+    embodied=Embodied( # (2)!
         gwp=GWP(),
         adpe=ADPe(),
         pe=PE(),
     ),
-    warnings=None,  # (4)!
+    warnings=None, # (3)!
     errors=None
 )
 ```
 
-1. Total impacts for all phases.
-2. Usage impacts for the electricity consumption impacts. Note that the energy is equal to the "total" energy impact.
-3. Embodied impacts for resource extract, manufacturing and transportation of hardware components allocated to the request. 
-4. List of [`WarningMessage`][status_messages.WarningMessage] and [`ErrorMessage`][status_messages.ErrorMessage].
+1. Usage impacts for the electricity consumption impacts. Note that the energy is equal to the "total" energy impact.
+2. Embodied impacts for resource extract, manufacturing and transportation of hardware components allocated to the request. 
+3. List of [`WarningMessage`][status_messages.WarningMessage] and [`ErrorMessage`][status_messages.ErrorMessage].
 
 
-You can extract an impact with:
+### Example of an impact value
 
-```python
->>> response.impacts.usage.gwp.value  # (1)!
-0.34    # Expressed in kgCO2eq.
-```
-
-1. Assuming you have made an inference and get the response in an `response` object.
-
-Or you could get **value range** impact instead:
+The impact objects named [`Energy`][impacts.modeling.Energy], [`GWP`][impacts.modeling.GWP], [`ADPe`][impacts.modeling.ADPe] or [`PE`][impacts.modeling.PE] all share the following structure:
 
 ```python
[…]
