+++ mkdocs.yml
@@ -129,7 +129,9 @@ markdown_extensions:
       line_spans: __span
       pygments_lang_class: true
   - pymdownx.inlinehilite
-  - pymdownx.snippets
+  - pymdownx.snippets:
+      auto_append:
+        - docs/_abbreviations
   - pymdownx.superfences
   - pymdownx.details
   - pymdownx.tabbed:
