### 2024-12-19 1f6ae5a
docs: corrections on environmental impacts
