### 2024-12-19 db9fe74
docs: corrections on environmental impacts
