### 2024-12-19 9b1ca60
docs: corrections on proprietary models
