### 2024-12-19 497bad0
docs: drafting hypotheses and limitations
