### 2024-12-19 824fe29
docs: drafting hypotheses and limitations
