+++ docs/methodology/index.md
@@ -22,7 +22,7 @@ Our assessment considers **three key environmental criteria**:
 - Abiotic Resource Depletion for Elements (ADPe): Assesses the consumption of raw minerals and metals, expressed in antimony equivalents.
 - Primary Energy (PE): Calculates energy consumed from natural sources, expressed in megajoules.
 
-Using a **bottom-up modeling approach**, we assess and aggregate the environmental impacts of all individual service components. This method differs from top-down approaches by allowing precise allocation of each resource's impact to the overall environmental footprint. The key advantage of bottom-up modeling is that our methodology can be customized for each provider that share information.
+Using a **bottom-up modeling approach**, we assess and aggregate the environmental impacts of all individual service components within scope. This method differs from top-down approaches by allowing precise allocation of each resource's impact to the overall environmental footprint. The key advantage of bottom-up modeling is that our methodology can be customized for each provider that share information.
 
 The methodology is grounded in **transparency** and **reproducibility**, utilizing open market and technical data to ensure our results are reliable and verifiable.
 
@@ -40,17 +40,58 @@ Because evaluating the environmental footprint of GenAI services is hard we make
 
 ## Hypotheses and limitations
 
-[//]: # (TODO: explain hardware and server)
+To be able to estimate environmental impacts of generative AI services at inference we took the approach of modeling the key components that compose the service. In this section we will list major assumptions we make when modeling environmental impacts as well as known limitations. When possible we will try to quantify the potential inaccuracies.
 
-[//]: # (TODO: explain data centers PUE + inaccuracies)
+!!! note "The hypotheses and limitations listed in this section are intentionally high-level. To get the full details you will need to dive into the methodology."
 
-[//]: # (TODO: explain electricity mix and wordwide average + inaccuracies)
+### On models
 
-[//]: # (TODO: explain embodied impacts and boaviztapi + inaccuracies)
+Two major information we are looking for is the required infrastructure to host the AI model, such as the number of GPUs as well as the energy consumption that results from doing an inference on the model.
 
-[//]: # (TODO: explain ai models benchmarks + inaccuracies)
+Assuming the **required infrastructure** for open models can be relatively straightforward because the model size is known. But for proprietary models this can very be challenging given that some AI provider do not disclose any technical information on that matter. That's why we rely on estimations of parameters count for closed models, to learn more [read the dedicated section](proprietary_models.md).
 
+Assuming the **energy consumption** for AI models is done through benchmarking open models. We tend to rely on external sources for benchmarking, but we conduct our own experiments as well. Because of our limited capacity and the technical complexity to host very big AI models we extrapolate the consumption of smaller models to bigger models.
 
+**Limitations:**
+
+* We do not account for other inference optimizations such as flash attention or batching.
+* We do not benchmark models bigger than 70 billion parameters.
+* We do not have benchmarks for multi-GPU deployments.
+* We do not account for the multiple modalities of a model.
+
+### On hardware
+
+We estimate the **required infrastructure** to run the service in terms of hardware. We consider that the service is hosted in the cloud on servers equipped with high-end GPUs. 
+
+**Limitations:**
+
+* We do not account for TPUs or other type of accelerators.
+* We do not account for networking or storage primitives.
+* We do not account for infrastructure overheads or utilization factors.
+
+### On data centers
+
+The type of services we model rely on high-end hardware that we consider is hosted by cloud service providers. Thus, we model data centers impacts as well and especially the overhead for cooling equipments.
+
[…]
