### 2024-12-20 21c810e
feat: include confidence intervals for linear regression
