+++ ecologits/impacts/llm.py
@@ -10,8 +10,11 @@ MODEL_QUANTIZATION_BITS = 4
 
 GPU_ENERGY_ALPHA = 8.91e-8
 GPU_ENERGY_BETA = 1.43e-6
+GPU_ENERGY_STDEV = 5.19e-7
 GPU_LATENCY_ALPHA = 8.02e-4
 GPU_LATENCY_BETA = 2.23e-2
+GPU_LATENCY_STDEV = 7.00e-6
+
 GPU_MEMORY = 80  # GB
 GPU_EMBODIED_IMPACT_GWP = 143
 GPU_EMBODIED_IMPACT_ADPE = 5.1e-3
@@ -35,8 +38,9 @@ def gpu_energy(
         model_active_parameter_count: float,
         output_token_count: float,
         gpu_energy_alpha: float,
-        gpu_energy_beta: float
-) -> float:
+        gpu_energy_beta: float,
+        gpu_energy_stdev: float
+) -> ValueOrRange:
     """
     Compute energy consumption of a single GPU.
 
@@ -45,12 +49,15 @@ def gpu_energy(
         output_token_count: Number of generated tokens.
         gpu_energy_alpha: Alpha parameter of the GPU linear power consumption profile.
         gpu_energy_beta: Beta parameter of the GPU linear power consumption profile.
+        gpu_energy_stdev: Standard deviation of the GPU linear power consumption profile.
 
     Returns:
-        The energy consumption of a single GPU in kWh.
+        The 95% confidence interval of energy consumption of a single GPU in kWh.
     """
-    return output_token_count * (gpu_energy_alpha * model_active_parameter_count + gpu_energy_beta)
-
+    gpu_energy_per_token_mean = gpu_energy_alpha * model_active_parameter_count + gpu_energy_beta
+    gpu_energy_min = output_token_count * (gpu_energy_per_token_mean - 1.96 * gpu_energy_stdev)
+    gpu_energy_max = output_token_count * (gpu_energy_per_token_mean + 1.96 * gpu_energy_stdev)
+    return RangeValue(min=max(0, gpu_energy_min), max=gpu_energy_max)
 
 @dag.asset
 def generation_latency(
@@ -58,8 +65,9 @@ def generation_latency(
         output_token_count: float,
         gpu_latency_alpha: float,
         gpu_latency_beta: float,
+        gpu_latency_stdev: float,
         request_latency: float,
-) -> float:
+) -> ValueOrRange:
     """
     Compute the token generation latency in seconds.
 
@@ -68,14 +76,19 @@ def generation_latency(
         output_token_count: Number of generated tokens.
         gpu_latency_alpha: Alpha parameter of the GPU linear latency profile.
         gpu_latency_beta: Beta parameter of the GPU linear latency profile.
+        gpu_energy_stdev: Standard deviation of the GPU linear latency profile.
         request_latency: Measured request latency (upper bound) in seconds.
 
     Returns:
         The token generation latency in seconds.
     """
-    gpu_latency = output_token_count * (gpu_latency_alpha * model_active_parameter_count + gpu_latency_beta)
-    return min(gpu_latency, request_latency)
-
+    gpu_latency_per_token_mean = gpu_latency_alpha * model_active_parameter_count + gpu_latency_beta
+    gpu_latency_min = output_token_count * (gpu_latency_per_token_mean - 1.96 * gpu_latency_stdev)
+    gpu_latency_max = output_token_count * (gpu_latency_per_token_mean + 1.96 * gpu_latency_stdev)
+    gpu_latency_interval = RangeValue(min=max(0, gpu_latency_min), max=gpu_latency_max)
+    if gpu_latency_interval < request_latency:
[…]
