### 2024-12-20 2b95315
fix: move range value tests in right file
