### 2024-12-20 76464df
chore: clean compute_llm_impacts and ignore PLR0912
