+++ ecologits/impacts/llm.py
@@ -440,7 +440,7 @@ def compute_llm_impacts_dag(
     return results
 
 
-def compute_llm_impacts(
+def compute_llm_impacts( # noqa: PLR0912
         model_active_parameter_count: ValueOrRange,
         model_total_parameter_count: ValueOrRange,
         output_token_count: float,
@@ -500,17 +500,19 @@ def compute_llm_impacts(
         for field in fields:
             if field in results:
                 min_result = results[field]
-                if isinstance(min_result, RangeValue):
+                max_result = res[field]
+                if isinstance(min_result, RangeValue) and isinstance(max_result, RangeValue):
                     min_value = cast(Union[float, int], min_result.min)
-                    max_result = res[field]
-                    if isinstance(max_result, RangeValue):
-                        max_value = cast(Union[float, int], max_result.max)
-                        results[field] = RangeValue(min=min_value, max=max_value)
-                    else:
-                        raise TypeError("Unexpected behaviour. With different parameters, DAG does not return the same type")
-                else:
+                    max_value = cast(Union[float, int], max_result.max)
+                    results[field] = RangeValue(min=min_value, max=max_value)
+                elif isinstance(min_result, (float, int)):
                     min_value = cast(Union[float, int], min_result)
-                    results[field] = RangeValue(min=min_value, max=res[field])
+                    max_value = cast(Union[float, int], max_result)
+                    results[field] = RangeValue(min=min_value, max=max_value)
+                else:
+                    raise TypeError(
+                        "Unexpected behaviour. With different parameters, DAG does not return the same type"
+                    )
             else:
                 results[field] = res[field]
 
