### 2024-12-20 7627143
test: add add, mul and truediv tests for RangeValue
