### 2025-01-12 004ec91
docs: reorganize hypotheses and limitations
