+++ docs/methodology/index.md
@@ -40,58 +40,9 @@ Because evaluating the environmental footprint of GenAI services is hard we make
 
 ## Hypotheses and limitations
 
-To be able to estimate environmental impacts of generative AI services at inference we took the approach of modeling the key components that compose the service. In this section we will list major assumptions we make when modeling environmental impacts as well as known limitations. When possible we will try to quantify the potential inaccuracies.
+Estimating the environmental impacts of generative AI services at inference can be really challenging because of the **lack of open data and transparent information** from the key players (AI/cloud providers, hardware manufacturers, environmental impact databases, etc.) In the [LLM inference](llm_inference.md) methodology we explain at high-level all the [**hypotheses and limitations**](llm_inference.md#hypothesis-and-limitations) of our bottom-up approach.
 
-!!! note "The hypotheses and limitations listed in this section are intentionally high-level. To get the full details you will need to dive into the methodology."
-
-### On models
-
-Two major information we are looking for is the required infrastructure to host the AI model, such as the number of GPUs as well as the energy consumption that results from doing an inference on the model.
-
-Assuming the **required infrastructure** for open models can be relatively straightforward because the model size is known. But for proprietary models this can very be challenging given that some AI provider do not disclose any technical information on that matter. That's why we rely on estimations of parameters count for closed models, to learn more [read the dedicated section](proprietary_models.md).
-
-Assuming the **energy consumption** for AI models is done through benchmarking open models. We tend to rely on external sources for benchmarking, but we conduct our own experiments as well. Because of our limited capacity and the technical complexity to host very big AI models we extrapolate the consumption of smaller models to bigger models.
-
-**Limitations:**
-
-* We do not account for other inference optimizations such as flash attention or batching.
-* We do not benchmark models bigger than 70 billion parameters.
-* We do not have benchmarks for multi-GPU deployments.
-* We do not account for the multiple modalities of a model.
-
-### On hardware
-
-We estimate the **required infrastructure** to run the service in terms of hardware. We consider that the service is hosted in the cloud on servers equipped with high-end GPUs. 
-
-**Limitations:**
-
-* We do not account for TPUs or other type of accelerators.
-* We do not account for networking or storage primitives.
-* We do not account for infrastructure overheads or utilization factors.
-
-### On data centers
-
-The type of services we model rely on high-end hardware that we consider is hosted by cloud service providers. Thus, we model data centers impacts as well and especially the overhead for cooling equipments.
-
-We consider the **Power Usage Effectiveness** (PUE) metric from data centers. These values can be quite complicated to get from the providers themselves. A good amount of data is available for providers that build their own data centers (such as hyperscalers). But part of the AI workloads are also located in non-hyperscale data centers or in co-located data centers. That's why we prefer to rely on a global average for PUE that can be overridden for providers that disclose more precise data.
-
-**Limitations:**
-
-* We do not account for the specific infrastructure or way to cooldown servers in data centers.
-* We do not account for the overhead of the cloud provider for internal services like backing up or monitoring.
-
-### On electricity mix
-
-Because getting precise information on where the consumption happens and in which data center we also rely on global averages for the environmental impacts of electricity consumption by cloud providers. By default, we use a worldwide value, and have country specific values that can be customized by the end-user or when the provider disclose more precise data on AI workload location.
-
-**Limitations:**
-
-* We do not account for local electricity generation for data centers, whether it be renewable energy (e.g. solar panels) or fossil fuels (e.g. diesel generators) 
-* We do not account for regional electricity mixes the smallest supported zone is a country.
-
-### On embodied impacts
-
[…]
