+++ docs/methodology/llm_inference.md
@@ -259,22 +259,77 @@ $$
 
 ## Hypothesis and limitations
 
-**Infrastructure and hardware:**
+To be able to estimate environmental impacts of LLMs at inference we took the approach of modeling the key components that compose the service. In this section we will list major assumptions we make when modeling environmental impacts as well as known limitations. When possible we will try to quantify the potential inaccuracies.
 
-- GPU: NVIDIA A100 80GB
-- Server (cloud instance): AWS p4de.24xlarge
-- Power Usage Effectiveness: 1.2 (arbitrary value for hyperscalers)
+### On models
 
-**AI model deployments:**
+Two major information we are looking for is the required infrastructure to host the AI model, such as the number of GPUs as well as the energy consumption that results from doing an inference on the model.
 
-- Backend: pytorch
-- Quantization: int4
-- Optimizations: none
-- Parallelism: none
+Assuming the **required infrastructure** for open models can be relatively straightforward because the model size is known. But for proprietary models this can very be challenging given that some AI provider do not disclose any technical information on that matter. That's why we rely on estimations of parameters count for closed models, to learn more [read the dedicated section](proprietary_models.md).
 
-**Impact factors:**
+Assuming the **energy consumption** for AI models is done through benchmarking open models. We tend to rely on external sources for benchmarking, but we conduct our own experiments as well. Because of our limited capacity and the technical complexity to host very big AI models we extrapolate the consumption of smaller models to bigger models.
 
-- Electricity mix from ADEME (default: World)
+**Hypotheses:**
+
+* Models are deployed with pytorch backend.
+* Models are quantized to 4 bits.
+
+**Limitations:**
+
+* We do not account for other inference optimizations such as flash attention, batching or parallelism.
+* We do not benchmark models bigger than 70 billion parameters.
+* We do not have benchmarks for multi-GPU deployments.
+* We do not account for the multiple modalities of a model (only text-to-text generation).
+
+### On hardware
+
+We estimate the **required infrastructure** to run the service in terms of hardware. We consider that the service is hosted in the cloud on servers equipped with high-end GPUs. 
+
+**Hypotheses:**
+
+* Models are deployed on NVIDIA A100 GPUs with 80GB of memory.
+* Base servers are similar to p4de.24xlarge AWS cloud instances.
+
+**Limitations:**
+
+* We do not account for TPUs or other type of accelerators.
+* We do not account for networking or storage primitives.
+* We do not account for infrastructure overheads or utilization factors.
+
+### On data centers
+
+The type of services we model rely on high-end hardware that we consider is hosted by cloud service providers. Thus, we model data centers impacts as well and especially the overhead for cooling equipments.
+
+We consider the **Power Usage Effectiveness** (PUE) metric from data centers. These values can be quite complicated to get from the providers themselves. A good amount of data is available for providers that build their own data centers (such as hyperscalers). But part of the AI workloads are also located in non-hyperscale data centers or in co-located data centers. That's why we prefer to rely on a global average for PUE that can be overridden for providers that disclose more precise data.
+
+**Hypotheses:**
+
+* PUE = 1.2 (arbitrary value for hyperscalers)
+
+**Limitations:**
+
+* We do not know precisely where are located the data centers that run AI models.
+* We do not account for the specific infrastructure or way to cooldown servers in data centers.
+* We do not account for the local electricity generation (private power plants) specific to the data center.
+* We do not account for the overhead of the cloud provider for internal services like backing up or monitoring.
+
+
+### On impact factors
+
+To transform physical values such as energy consumption into environmental impacts we use **impact factors**. These can be hard to estimate and precise and up-to-date data is rarely open to use.
+
+**Hypotheses:**
+
[…]
