### 2025-01-12 cd285fe
docs: reorganize hypotheses and limitations
