### 2025-01-12 2fca506
docs: reorganize hypotheses and limitations
