### 2025-01-15 5d3a8cd
feat: WIP trying a few things before config dev
