+++ ecologits/tracers/utils.py
@@ -1,64 +1,13 @@
-from typing import Optional
-
-from pydantic import BaseModel
-
+from ecologits._ecologits import EcoLogits
+from ecologits.domain import ImpactsOutput
 from ecologits.electricity_mix_repository import electricity_mixes
 from ecologits.impacts.llm import compute_llm_impacts
-from ecologits.impacts.modeling import GWP, PE, ADPe, Embodied, Energy, Usage
 from ecologits.log import logger
 from ecologits.model_repository import ParametersMoE, models
-from ecologits.status_messages import ErrorMessage, ModelNotRegisteredError, WarningMessage, ZoneNotRegisteredError
+from ecologits.status_messages import ModelNotRegisteredError, ZoneNotRegisteredError
 
 
-class ImpactsOutput(BaseModel):
-    """
-    Impacts output data model.
-
-    Attributes:
-        energy: Total energy consumption
-        gwp: Total Global Warming Potential (GWP) impact
-        adpe: Total Abiotic Depletion Potential for Elements (ADPe) impact
-        pe: Total Primary Energy (PE) impact
-        usage: Impacts for the usage phase
-        embodied: Impacts for the embodied phase
-        warnings: List of warnings
-        errors: List of errors
-    """
-    energy: Optional[Energy] = None
-    gwp: Optional[GWP] = None
-    adpe: Optional[ADPe] = None
-    pe: Optional[PE] = None
-    usage: Optional[Usage] = None
-    embodied: Optional[Embodied] = None
-    warnings: Optional[list[WarningMessage]] = None
-    errors: Optional[list[ErrorMessage]] = None
-
-    @property
-    def has_warnings(self) -> bool:
-        return isinstance(self.warnings, list) and len(self.warnings) > 0
-
-    @property
-    def has_errors(self) -> bool:
-        return isinstance(self.errors, list) and len(self.errors) > 0
-
-    def add_warning(self, warning: WarningMessage) -> None:
-        if self.warnings is None:
-            self.warnings = []
-        self.warnings.append(warning)
-
-    def add_errors(self, error: ErrorMessage) -> None:
-        if self.errors is None:
-            self.errors = []
-        self.errors.append(error)
-
-
-def llm_impacts(
-    provider: str,
-    model_name: str,
-    output_token_count: int,
-    request_latency: float,
-    electricity_mix_zone: str = "WOR",
-) -> ImpactsOutput:
+def llm_impacts(provider: str, model_name: str, output_token_count: int, request_latency: float) -> ImpactsOutput:
     """
     High-level function to compute the impacts of an LLM generation request.
 
@@ -72,7 +21,9 @@ def llm_impacts(
     Returns:
         The impacts of an LLM generation request.
     """
-
+    electricity_mix_zone: str = EcoLogits.config.providers_electricity_mixes_zones.get(
+        provider, EcoLogits.config.electricity_mix_zone if not None else "WOR"
+    )
     model = models.find_model(provider=provider, model_name=model_name)
     if model is None:
         error = ModelNotRegisteredError(message=f"Could not find model `{model_name}` for {provider} provider.")
@@ -92,9 +43,9 @@ def llm_impacts(
         logger.warning_once(str(error))
         return ImpactsOutput(errors=[error])
 
-    if_electricity_mix_adpe=electricity_mix.adpe
-    if_electricity_mix_pe=electricity_mix.pe
-    if_electricity_mix_gwp=electricity_mix.gwp
+    if_electricity_mix_adpe = electricity_mix.adpe
+    if_electricity_mix_pe = electricity_mix.pe
+    if_electricity_mix_gwp = electricity_mix.gwp
     impacts = compute_llm_impacts(
         model_active_parameter_count=model_active_params,
         model_total_parameter_count=model_total_params,
