+++ tests/conftest.py
@@ -3,7 +3,6 @@ import os
 import pytest
 import tiktoken
 
-from ecologits import EcoLogits
 
 
 tiktoken.get_encoding("cl100k_base")
@@ -37,6 +36,4 @@ def vcr_config():
     ]}
 
 
-@pytest.fixture(scope="session")
-def tracer_init():
-    EcoLogits.init()
+
