+++ ecologits/model_repository.py
@@ -115,4 +115,5 @@ class ModelRepository:
         return cls(models=model_list, aliases=alias_list)
 
 
+
 models = ModelRepository.from_json()

[…]
