### 2025-01-19 e4d14f2
feat: added config and tests. Todo: update docs
