### 2025-02-02 a48e3c4
feat: added docs and updated conf handling
