### 2025-02-05 1786e4c
add confidence intervals to docs
