+++ .github/workflows/pre-commit.yaml
@@ -15,7 +15,6 @@ jobs:
     - name: Install poetry
       run: |
         python -m pip install poetry
-        poetry export --with dev --format=requirements.txt --output=requirements.txt
     - name: set PY
       run: echo "PY=$(python -c 'import hashlib, sys;print(hashlib.sha256(sys.version.encode()+sys.executable.encode()).hexdigest())')" >> $GITHUB_ENV
     - uses: actions/cache@v3
