### 2025-02-05 d14b30e
fix: 3.0.0 -> 3.0.1
