+++ .github/workflows/pre-commit.yaml
@@ -21,4 +21,4 @@ jobs:
       with:
         path: ~/.cache/pre-commit
         key: pre-commit|${{ env.PY }}|${{ hashFiles('.pre-commit-config.yaml') }}
-    - uses: pre-commit/action@v3.0.0
+    - uses: pre-commit/action@v3.0.1
