+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix]
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.3.0
+    rev: v5.0.0
     hooks:
     - id: check-merge-conflict
     - id: mixed-line-ending
