+++ .pre-commit-config.yaml
@@ -16,7 +16,7 @@ repos:
   #  - id: bandit
   #    exclude: tests/
   - repo: https://github.com/Lucas-C/pre-commit-hooks-safety
-    rev: v1.3.1
+    rev: v1.4.0
     hooks:
     - id: python-safety-dependencies-check
   - repo: https://github.com/pre-commit/mirrors-mypy
