### 2025-02-05 501b522
fix: update both pre-commit and safety hooks
