### 2025-02-05 1ac31d5
fix: temporarily remove safety checks
