### 2025-02-05 3d56048
fix: update anthropic, fixes anthropic tests
