+++ pyproject.toml
@@ -41,7 +41,7 @@ pydantic = ">=2,<3"
 packaging = ">=23.2,<25.0"
 openai = { version = "^1.12.0", optional = true }
 mistralai = { version = ">=0.4.3,<2.0.0", optional = true }
-anthropic = { version = "^0.29.0", optional = true }
+anthropic = { version = "^0.40.0", optional = true }
 cohere = {version = "^5.2.5", optional = true}
 huggingface-hub = { version = "^0.22.2",  optional = true }
 tiktoken = { version = "^0.8.0", optional = true }
