+++ .pre-commit-config.yaml
@@ -15,10 +15,10 @@ repos:
   #  hooks:
   #  - id: bandit
   #    exclude: tests/
-  # - repo: https://github.com/Lucas-C/pre-commit-hooks-safety
-  #   rev: v1.4.0
-  #   hooks:
-  #     - id: python-safety-dependencies-check
+  - repo: https://github.com/Lucas-C/pre-commit-hooks-safety
+    rev: v1.4.0
+    hooks:
+      - id: python-safety-dependencies-check
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: 'v1.13.0'
     hooks:
