### 2025-02-06 bae2bca
chore: migrate to poetry 2.0
