### 2025-02-06 58fefde
fix: try lock after installing export
