+++ .github/workflows/pre-commit.yaml
@@ -16,6 +16,7 @@ jobs:
       run: |
         python -m pip install poetry
         python -m pip install poetry-plugin-export
+        poetry lock
         poetry export --with dev --format=requirements.txt --output=requirements.txt
     - name: set PY
       run: echo "PY=$(python -c 'import hashlib, sys;print(hashlib.sha256(sys.version.encode()+sys.executable.encode()).hexdigest())')" >> $GITHUB_ENV
