### 2025-02-06 81cc947
fix: remove litellm from optional
