+++ pyproject.toml
@@ -35,27 +35,21 @@ classifiers = [
 ]
 requires-python = ">=3.9,<4"
 dependencies = [
-    "wrapt (>=1.17.2,<2.0.0)",
+    "wrapt (>=1.14.1,<2.0.0)",
     "pydantic (>=2.10.6,<3.0.0)",
     "packaging (>=24.2,<25.0)",
+    "litellm (>=1.60.5,<2.0.0)",
+    "rapidfuzz (>=3.12.1,<4.0.0)"
 ]
 
 [project.optional-dependencies]
-openai=["openai (>=1.61.1,<2.0.0)"]
-mistralai=["mistralai (>=1.5.0,<2.0.0)"]
-anthropic=["anthropic (>=0.45.2,<0.46.0)"]
-cohere=["cohere (>=5.13.12,<6.0.0)"]
-huggingface-hub=[
-    "huggingface-hub (>=0.28.1,<0.29.0)",
-    "tiktoken (>=0.8.0,<0.9.0)",
-    "aiohttp (>=3.11.12,<4.0.0)",
-    "minijinja (>=2.7.0,<3.0.0)",
-]
-google-generativeai=["google-generativeai (>=0.8.4,<0.9.0)"]
-litellm=[
-    "litellm (>=1.60.5,<2.0.0)",
-    "rapidfuzz (>=3.12.1,<4.0.0)",
-]
+openai = ["openai (>=1.61.1,<2.0.0)"]
+mistralai = ["mistralai (>=1.5.0,<2.0.0)"]
+anthropic = ["anthropic (>=0.45.2,<0.46.0)"]
+cohere = ["cohere (>=5.13.12,<6.0.0)"]
+huggingface-hub = ["huggingface-hub (>=0.28.1,<0.29.0)", "tiktoken (>=0.8.0,<0.9.0)", "aiohttp (>=3.11.12,<4.0.0)", "minijinja (>=2.7.0,<3.0.0)"]
+google-generativeai = ["google-generativeai (>=0.8.4,<0.9.0)"]
+
 
 
 [tool.poetry]
