### 2025-02-06 57eb0c4
fix: remove azure tests
