+++ tests/cassettes/test_openai/test_azure_openai_async_chat.yaml
@@ -1,79 +1,0 @@
-interactions:
-- request:
-    body: '{"messages": [{"role": "user", "content": "Hello World!"}], "model": "gpt-4o-mini"}'
-    headers:
-      accept:
-      - application/json
-      accept-encoding:
-      - gzip, deflate
-      connection:
-      - keep-alive
-      content-length:
-      - '83'
-      content-type:
-      - application/json
-      host:
-      - ecologits-test.openai.azure.com
-      user-agent:
-      - AsyncAzureOpenAI/Python 1.35.14
-      x-stainless-arch:
-      - x64
-      x-stainless-async:
-      - async:asyncio
-      x-stainless-lang:
-      - python
-      x-stainless-os:
-      - MacOS
-      x-stainless-package-version:
-      - 1.35.14
-      x-stainless-runtime:
-      - CPython
-      x-stainless-runtime-version:
-      - 3.9.18
-    method: POST
-    uri: https://ecologits-test.openai.azure.com//openai/deployments/gpt-4o-mini/chat/completions?api-version=2024-06-01
-  response:
-    body:
-      string: '{"choices":[{"content_filter_results":{"hate":{"filtered":false,"severity":"safe"},"protected_material_code":{"filtered":false,"detected":false},"protected_material_text":{"filtered":false,"detected":false},"self_harm":{"filtered":false,"severity":"safe"},"sexual":{"filtered":false,"severity":"safe"},"violence":{"filtered":false,"severity":"safe"}},"finish_reason":"stop","index":0,"logprobs":null,"message":{"content":"Hello!
-        How can I assist you today?","role":"assistant"}}],"created":1724436865,"id":"chatcmpl-9zT5lrgOwZH8PlqWOpt7WkIeSqiDX","model":"gpt-4o-mini","object":"chat.completion","prompt_filter_results":[{"prompt_index":0,"content_filter_results":{"hate":{"filtered":false,"severity":"safe"},"jailbreak":{"filtered":false,"detected":false},"self_harm":{"filtered":false,"severity":"safe"},"sexual":{"filtered":false,"severity":"safe"},"violence":{"filtered":false,"severity":"safe"}}}],"system_fingerprint":"fp_80a1bad4c7","usage":{"completion_tokens":9,"prompt_tokens":10,"total_tokens":19}}
-
-        '
-    headers:
-      Content-Length:
-      - '1010'
-      Content-Type:
-      - application/json
-      Date:
-      - Fri, 23 Aug 2024 18:14:25 GMT
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      apim-request-id:
-      - e21a93af-792a-4126-b587-bcd52def1b75
-      azureml-model-session:
-      - d001-20240820051422
-      cmp-upstream-response-duration:
-      - '148'
-      ms-azureml-model-time:
-      - '185'
-      x-accel-buffering:
-      - 'no'
-      x-content-type-options:
-      - nosniff
-      x-envoy-upstream-service-time:
-      - '191'
-      x-ms-client-request-id:
-      - e21a93af-792a-4126-b587-bcd52def1b75
-      x-ms-rai-invoked:
-      - 'true'
-      x-ms-region:
-      - East US
-      x-ratelimit-remaining-requests:
-      - '9'
-      x-ratelimit-remaining-tokens:
-      - '356'
-      x-request-id:
-      - 5b8caf9d-4476-4bc9-a7f4-27a5cb83534f
-    status:
-      code: 200
-      message: OK
-version: 1
