### 2025-02-06 5bcd6cd
fix: remove azure tests
