### 2025-02-06 a7c3e0d
fix: azure tests
