+++ tests/test_openai.py
@@ -52,7 +52,6 @@ async def test_openai_async_stream_chat(tracer_init):
 
 
 @pytest.mark.vcr
-@pytest.mark.skip(reason="Waiting for a new testing instance")
 def test_azure_openai_chat(tracer_init):
     client = AzureOpenAI(azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"))
     response = client.chat.completions.create(        
@@ -65,7 +64,6 @@ def test_azure_openai_chat(tracer_init):
 
 @pytest.mark.vcr
 @pytest.mark.asyncio
-@pytest.mark.skip(reason="Waiting for a new testing instance")
 async def test_azure_openai_async_chat(tracer_init):
     client = AsyncAzureOpenAI(azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"))
     response = await client.chat.completions.create(
@@ -77,7 +75,6 @@ async def test_azure_openai_async_chat(tracer_init):
 
 
 @pytest.mark.vcr
-@pytest.mark.skip(reason="Waiting for a new testing instance")
 def test_azure_openai_stream_chat(tracer_init):
     client = AzureOpenAI(azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"))
     stream = client.chat.completions.create(
@@ -91,7 +88,6 @@ def test_azure_openai_stream_chat(tracer_init):
 
 @pytest.mark.vcr
 @pytest.mark.asyncio
-@pytest.mark.skip(reason="Waiting for a new testing instance")
 async def test_azure_openai_async_stream_chat(tracer_init):
     client = AsyncAzureOpenAI(azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"))
     stream = await client.chat.completions.create(
