### 2025-02-06 b26f867
fix: azure tests
