+++ tests/conftest.py
@@ -18,7 +18,7 @@ def environment():
     set_envvar_if_unset("GOOGLE_API_KEY", "test-api-key")
     set_envvar_if_unset("AZURE_OPENAI_API_KEY", "test-api-key")
     set_envvar_if_unset("AZURE_OPENAI_ENDPOINT", "https://ecologits-azure-openai.openai.azure.com/openai/deployments/gpt-4o-mini/chat/completions?api-version=2024-08-01-preview")
-    set_envvar_if_unset("OPENAI_API_VERSION", "2024-10-21")
+    set_envvar_if_unset("OPENAI_API_VERSION", "2024-06-01")
     set_envvar_if_unset("AZURE_MODEL_DEPLOYMENT", "gpt-4o-mini")
 
 
