+++ .github/workflows/pytest.yaml
@@ -17,8 +17,6 @@ jobs:
       - name: Set up python
         id: setup-python
         uses: actions/setup-python@v5
-        with:
-          python-version: "3.9"
       #----------------------------------------------
       #  -----  install & configure poetry  -----
       #----------------------------------------------
