### 2025-02-07 687a9d4
fix: remove litellm from deps, add requests, httpx, tqdm
