+++ pyproject.toml
@@ -36,11 +36,12 @@ classifiers = [
 requires-python = ">=3.9,<4"
 dependencies = [
     "wrapt (>=1.14.1,<2.0.0)",
-    "pydantic (>=2.10.6,<3.0.0)",
+    "pydantic >=1.9.0,<3",
     "packaging (>=24.2,<25.0)",
-    "litellm (>=1.60.5,<2.0.0)",
-    "rapidfuzz (>=3.12.1,<4.0.0)"
-]
+    "requests >=2.26.0",            # only needed for `poetry lock` to work with `poetry >2.0`
+    "httpx >=0.23.0,<1",            # only needed for `poetry lock` to work with `poetry >2.0`
+    "tqdm >4",                      # only needed for `poetry lock` to work with `poetry >2.0`
+]   
 
 [project.optional-dependencies]
 openai = ["openai (>=1.61.1,<2.0.0)"]
@@ -49,7 +50,7 @@ anthropic = ["anthropic (>=0.45.2,<0.46.0)"]
 cohere = ["cohere (>=5.13.12,<6.0.0)"]
 huggingface-hub = ["huggingface-hub (>=0.28.1,<0.29.0)", "tiktoken (>=0.8.0,<0.9.0)", "aiohttp (>=3.11.12,<4.0.0)", "minijinja (>=2.7.0,<3.0.0)"]
 google-generativeai = ["google-generativeai (>=0.8.4,<0.9.0)"]
-
+litellm = ["litellm (>=1.60.6,<2.0.0)", "rapidfuzz (>=3.12.1,<4.0.0)"]
 
 
 [tool.poetry]
