+++ pyproject.toml
@@ -103,7 +103,7 @@ ignore_errors = true
 
 
 [tool.ruff]
-select = [
+lint.select = [
     "A",
     "ANN",
     "ARG",
@@ -135,10 +135,9 @@ select = [
     "YTT"
 ]
 
-ignore = [
+lint.ignore = [
     "A003",
-    "ANN101",
-    "ANN102",
+
     "ANN401",
     "N805",
     "N818",
@@ -150,7 +149,7 @@ ignore = [
     "TRY003"
 ]
 
-fixable = [
+lint.fixable = [
     "A",
     "ANN",
     "ARG",
@@ -180,7 +179,7 @@ fixable = [
     "W",
     "YTT"
 ]
-unfixable = []
+lint.unfixable = []
 
 exclude = [
     ".bzr",
@@ -205,12 +204,12 @@ exclude = [
 
 line-length = 120
 
-dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
 target-version = "py39"
 
-[tool.ruff.mccabe]
+[tool.ruff.lint.mccabe]
 max-complexity = 10
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "google"
