### 2025-02-09 de34f2e
Setting Electricity mix and providers from config file (#108)
* feat: added config and tests. Todo: update docs

* feat: added docs and updated conf handling

* chore: better with precommit

* feat: fix mypy

* feat: fix tests

* feat: fix mypy

* feat: update pre-commit

* feat: fix tests
