+++ tests/config/toml_with_no_ecologits.toml
@@ -1,0 +1,2 @@
+[other_tool]
+other_key = "othervalue"
