### 2025-02-12 5dbbfb9
Revert "Setting Electricity mix and providers from config file (#108)"
This reverts commit de34f2ebfdfe403543046669131b1c553738b8e7.
