+++ ecologits/__init__.py
@@ -2,6 +2,6 @@ from ._ecologits import EcoLogits
 
 __version__ = "0.5.2"
 __all__ = [
-    "EcoLogits",
-    "__version__"
+    "__version__",
+    "EcoLogits"
 ]
