+++ ecologits/tracers/huggingface_tracer.py
@@ -69,7 +69,9 @@ def huggingface_chat_wrapper_stream(
 ) -> Iterable[ChatCompletionStreamOutput]:
     timer_start = time.perf_counter()
     stream = wrapped(*args, **kwargs)
-    for token_count, chunk in enumerate(stream, start=1):
+    token_count = 0
+    for chunk in stream:
+        token_count += 1
         request_latency = time.perf_counter() - timer_start
         impacts = llm_impacts(
             provider=PROVIDER,
