+++ pyproject.toml
@@ -52,6 +52,7 @@ huggingface-hub = ["huggingface-hub (>=0.28.1,<0.29.0)", "tiktoken (>=0.8.0,<0.9
 google-generativeai = ["google-generativeai (>=0.8.4,<0.9.0)"]
 litellm = ["litellm (>=1.60.6,<2.0.0)", "rapidfuzz (>=3.12.1,<4.0.0)"]
 
+
 [tool.poetry]
 requires-poetry = ">=2.0"
 
@@ -103,7 +104,7 @@ ignore_errors = true
 
 
 [tool.ruff]
-lint.select = [
+select = [
     "A",
     "ANN",
     "ARG",
@@ -135,8 +136,10 @@ lint.select = [
     "YTT"
 ]
 
-lint.ignore = [
+ignore = [
     "A003",
+    "ANN101",
+    "ANN102",
     "ANN401",
     "N805",
     "N818",
@@ -148,7 +151,7 @@ lint.ignore = [
     "TRY003"
 ]
 
-lint.fixable = [
+fixable = [
     "A",
     "ANN",
     "ARG",
@@ -178,7 +181,7 @@ lint.fixable = [
     "W",
     "YTT"
 ]
-lint.unfixable = []
+unfixable = []
 
 exclude = [
     ".bzr",
@@ -203,12 +206,12 @@ exclude = [
 
 line-length = 120
 
-lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
+dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
 target-version = "py39"
 
-[tool.ruff.lint.mccabe]
+[tool.ruff.mccabe]
 max-complexity = 10
 
-[tool.ruff.lint.pydocstyle]
+[tool.ruff.pydocstyle]
 convention = "google"
