+++ docs/scripts/gen_references.py
@@ -26,5 +26,5 @@ for path in sorted(Path(PACKAGE).rglob("*.py")):
 
     mkdocs_gen_files.set_edit_path(full_doc_path, Path("../") / path)
 
-with mkdocs_gen_files.open("reference/SUMMARY.md", "w") as nav_file:
+with mkdocs_gen_files.open("reference/SUMMARY.md", "w") as nav_file:  #
     nav_file.writelines(nav.build_literate_nav())
