+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 repos:
   - repo: https://github.com/charliermarsh/ruff-pre-commit
     # Ruff version.
-    rev: 'v0.9.5'
+    rev: 'v0.0.254'
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix]
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v5.0.0
+    rev: v4.3.0
     hooks:
     - id: check-merge-conflict
     - id: mixed-line-ending
@@ -16,10 +16,10 @@ repos:
   #  - id: bandit
   #    exclude: tests/
   - repo: https://github.com/Lucas-C/pre-commit-hooks-safety
-    rev: v1.4.0
+    rev: v1.3.1
     hooks:
     - id: python-safety-dependencies-check
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: 'v1.15.0'
+    rev: 'v1.13.0'
     hooks:
       - id: mypy
