+++ tests/config/toml_with_no_ecologits.toml
@@ -1,2 +1,0 @@
-[other_tool]
-other_key = "othervalue"

[…]
