+++ .github/PR_TEMPLATE/pr_template.md
@@ -1,0 +1,8 @@
+My PR is adressing the following issue: 
+
+I made sure I did the relevant tasks among:
+
+- [ ] I detailed my objectives in an issue with the feature request or bug report
+- [ ] I ran pre-commit checks and unit tests
+- [ ] New lines of code are tested with relevant unit tests
+- [ ] My feature is documented in the package documentation and I made sure the docs build locally
