### 2025-02-13 ba53719
Confidence intervals from linear regression (#104)
* feat: include confidence intervals for linear regression

* clean: remove useless line

* fix: move range value tests in right file

* chore: clean compute_llm_impacts and ignore PLR0912

* test: add add, mul and truediv tests for RangeValue

* add confidence intervals to docs

* chore: poetry lock

* docs: clarifications review
