+++ tests/config/test_config.py
@@ -9,7 +9,7 @@ default_providers = providers=list(set(_INSTRUMENTS.keys()))
 default_electricity_mix = "WOR"
 
 user_electricity_mix = "FRA"
-user_providers_list = list(set(["openai", "mistral"]))
+user_providers_list = ["openai", "mistral"]
 user_single_provider = ["openai"]
 
 
@@ -20,7 +20,7 @@ class TestEcoLogitsConfig:
     def test_working_config_provider_list(self, toml):
         EcoLogits.init(config_path=default_path)
 
-        assert EcoLogits.config.providers == user_providers_list
+        assert EcoLogits.config.providers.sort() == user_providers_list.sort()
         assert EcoLogits.config.electricity_mix_zone == user_electricity_mix
 
     @patch(target="ecologits._ecologits.EcoLogits._read_ecologits_config", return_value = {"electricity_mix_zone":user_electricity_mix, "providers":user_single_provider})
@@ -34,48 +34,48 @@ class TestEcoLogitsConfig:
         with caplog.at_level(logging.WARNING):
             EcoLogits.init(config_path=default_path)
             
-        assert not (set(EcoLogits.config.providers) ^ set(default_providers))
+        assert EcoLogits.config.providers.sort() == user_providers_list.sort()
         assert EcoLogits.config.electricity_mix_zone == default_electricity_mix
-        assert "Provided file does not exist, will fall back on default values" in caplog.text
+        assert "File does not exist, falling back on defaults" in caplog.text
 
     @patch(target="ecologits._ecologits.EcoLogits._read_ecologits_config", return_value = {"electricity_mix_zone":user_electricity_mix})
     def test_only_elec_mix_provided(self, patch):
 
         EcoLogits.init(config_path=default_path)
-        assert not (set(EcoLogits.config.providers) ^ set(default_providers))
+        EcoLogits.config.providers.sort() == user_providers_list.sort()
         assert EcoLogits.config.electricity_mix_zone == user_electricity_mix
 
     @patch(target="ecologits._ecologits.EcoLogits._read_ecologits_config", return_value = {"providers":user_providers_list})
     def test_only_provider_in_config(self, patch):
         EcoLogits.init(config_path=default_path)
 
-        assert EcoLogits.config.providers == user_providers_list
+        EcoLogits.config.providers.sort() == user_providers_list.sort()
         assert EcoLogits.config.electricity_mix_zone == default_electricity_mix
 
     def test_no_ecologits_key_in_toml(self, caplog):
         with caplog.at_level(logging.WARNING):
             EcoLogits.init(config_path="./tests/config/toml_with_no_ecologits.toml")
 
-        assert not (set(EcoLogits.config.providers) ^ set(default_providers))
+        EcoLogits.config.providers.sort() == user_providers_list.sort()
         assert EcoLogits.config.electricity_mix_zone == default_electricity_mix
-        assert "Provided file did not contain the ecologits key. Falling back on default configuration" in caplog.text
+        assert "File does not have the 'ecologits' key, falling back on defaults" in caplog.text
 
     def test_init_parameters_both_provided(self):
         EcoLogits.init(providers = user_providers_list, electricity_mix_zone=user_electricity_mix)
 
-        assert EcoLogits.config.providers == user_providers_list
+        EcoLogits.config.providers.sort() == user_providers_list.sort()
         assert EcoLogits.config.electricity_mix_zone == user_electricity_mix
 
     def test_init_parameters_elec_only_provided(self):
         EcoLogits.init(electricity_mix_zone=user_electricity_mix)
 
-        assert not (set(EcoLogits.config.providers) ^ set(default_providers))
[…]
