+++ pyproject.toml
@@ -66,7 +66,6 @@ pytest-recording = "^0.13.1"
 pytest-dotenv = "^0.5.2"
 pytest-asyncio = "^0.23.6"
 numpy = "^2.0.0"
-toml = "^0.10.2"
 mypy = "^1.13.0"
 
 
