+++ tests/test_version.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import Optional
 
-import toml
+import tomllib
 
 from ecologits import __version__
 
@@ -13,6 +13,6 @@ def test_version_alignment():
 
 def get_poetry_version() -> Optional[str]:
     path = Path(__file__).resolve().parents[1] / 'pyproject.toml'
-    with open(str(path), "r") as fd:
-        pyproject = toml.loads(fd.read())
+    with open(str(path), "rb") as fd:
+        pyproject = tomllib.load(fd)
         return pyproject['project']['version']
