### 2025-02-15 f9cbf3f
docs: remove repetitive info about the scope
