### 2025-02-15 906a669
fix: add tomli dependency for python version < 3.11
