+++ tests/test_version.py
@@ -1,9 +1,8 @@
 from pathlib import Path
 from typing import Optional
 
-import tomllib
-
 from ecologits import __version__
+from ecologits.utils.toml import load_toml
 
 
 def test_version_alignment():
@@ -13,6 +12,5 @@ def test_version_alignment():
 
 def get_poetry_version() -> Optional[str]:
     path = Path(__file__).resolve().parents[1] / 'pyproject.toml'
-    with open(str(path), "rb") as fd:
-        pyproject = tomllib.load(fd)
-        return pyproject['project']['version']
+    pyproject = load_toml(str(path))
+    return pyproject['project']['version']
