+++ pyproject.toml
@@ -41,6 +41,7 @@ dependencies = [
     "requests >=2.26.0",            # only needed for `poetry lock` to work with `poetry >2.0`
     "httpx >=0.23.0,<1",            # only needed for `poetry lock` to work with `poetry >2.0`
     "tqdm >4",                      # only needed for `poetry lock` to work with `poetry >2.0`
+    "tomli (>=2.0.1,<3.0) ; python_version < '3.11'"
 ]   
 
 [project.optional-dependencies]
