### 2025-02-18 a59e5fe
docs: switch to assumption instead of hypothesis for consistency
