### 2025-02-18 f8f766f
docs: switch to assumption instead of hypothesis for consistency
