### 2025-02-18 a8d63b9
docs: add link to range value explanation in get started
