### 2025-02-18 b04dde8
docs: add precision on PUE value
