### 2025-02-19 032c447
docs: add benchmarking data section in assumptions and limitations
