+++ docs/methodology/llm_inference.md
@@ -279,6 +279,14 @@ Assuming the **energy consumption** for AI models is done through benchmarking o
 * We do not have benchmarks for multi-GPU deployments.
 * We do not account for the multiple modalities of a model (only text-to-text generation).
 
+### On benchmarking data
+
+We use linear regression models to approximate energy consumption and latency as a function of the number of active parameters in the LLM. We represent the linear model as $Y = a \times X + b + \epsilon$, where $Y$ is the predicted value, $X$ is the input variable, $a$ and $b$ are the regression coefficients, and $\epsilon$ is the error term. We rely on two main assumptions about the errors or residuals:
+
+1. Normally distributed errors: We assume that the errors ($\epsilon$) follow a normal distribution with a mean of zero and a constant variance, represented as $\epsilon \sim \mathcal{N}(0, \sigma^2)$. 
+
+2. 95% confidence interval: To account for uncertainty, we use the 95% confidence interval, calculated using the standard deviation of the errors (\sigma) and the 97.5th percentile point of the standard normal distribution (approximately 1.96).
+
 ### On hardware
 
 We estimate the **required infrastructure** to run the service in terms of hardware. We consider that the service is hosted in the cloud on servers equipped with high-end GPUs. 
@@ -323,7 +331,6 @@ To transform physical values such as energy consumption into environmental impac
 **Limitations:**
 
 * We do not account for local electricity generation for data center or regional electricity mixes the smallest supported zone is a country.
-
 
 ### On embodied impacts
 
