### 2025-02-19 6db87ed
docs: fix missing "per token"
