### 2025-02-25 4e887b8
docs: update assumption on benchmarking data
