### 2025-02-27 c23cd90
test: update matching of claude models for litellm
