### 2025-02-28 b9ef2a8
docs: add downloads per month on README.md
