### 2025-02-28 d4a91a1
chore: bump version to 0.6.0
