### 2025-02-28 fba5d67
chore: bump version to 0.6.1
