### 2025-03-10 9695cc6
fix: missing stream in acompletion
