### 2025-03-10 e5fbca7
chore: bump version to 0.6.2
