+++ ecologits/__init__.py
@@ -1,6 +1,6 @@
 from ._ecologits import EcoLogits
 
-__version__ = "0.6.1"
+__version__ = "0.6.2"
 __all__ = [
     "__version__",
     "EcoLogits"
