### 2025-03-31 4ef0ee6
doc: clarify number of parameters in billion
