### 2025-04-04 bf763f0
feat: add base homepage
