### 2025-04-04 00a4274
feat: add base homepage
