+++ docs/home.md
@@ -1,0 +1,9 @@
+---
+template: home.html
+title: EcoLogits
+hide:
+  - navigation
+  - toc
+---
+
+Welcome to EcoLogits.
