### 2025-04-04 6fd0826
chore: bump version to 0.6.3
