### 2025-04-07 41ef406
Fix llm_impacts().
