+++ ecologits/__init__.py
@@ -2,6 +2,6 @@ from ._ecologits import EcoLogits
 
 __version__ = "0.6.3"
 __all__ = [
-    "__version__",
-    "EcoLogits"
+    "EcoLogits",
+    "__version__"
 ]
