+++ pyproject.toml
@@ -66,7 +66,7 @@ requires-poetry = ">=2.0"
 optional = true
 
 [tool.poetry.group.dev.dependencies]
-pre-commit = "^2.20.0"
+pre-commit = "^4.2.0"
 pytest = "^7.2.0"
 # ruff = "^0.0.254"
 pytest-recording = "^0.13.1"
