### 2025-04-23 8a3a1fd
fix: add safety req
