+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 repos:
   - repo: https://github.com/charliermarsh/ruff-pre-commit
     # Ruff version.
-    rev: 'v0.0.254'
+    rev: 'v0.11.6'
     hooks:
       - id: ruff
         args: [--fix, --exit-non-zero-on-fix]
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.3.0
+    rev: v5.0.0
     hooks:
     - id: check-merge-conflict
     - id: mixed-line-ending
@@ -16,7 +16,7 @@ repos:
   #  - id: bandit
   #    exclude: tests/
   - repo: https://github.com/Lucas-C/pre-commit-hooks-safety
-    rev: v1.3.1
+    rev: v1.4.0
     hooks:
     - id: python-safety-dependencies-check
   - repo: https://github.com/pre-commit/mirrors-mypy
