### 2025-04-23 4c3e2ad
fix: upgrade safety hook + lock
