### 2025-04-23 2efe102
fix: upgrade safety hook + lock
