### 2025-04-29 fdc934f
update config counters
