### 2025-04-29 db9aeff
update config counters
