### 2025-04-29 a321f87
update config counters
