### 2025-05-08 837e776
chore: remove setup.py
