### 2025-05-08 a9edfa6
docs: remove example docs for otel in README.md
Will add a dedicated section in the docs later.
