### 2025-05-08 9762814
chore: add otel dependencies + update
