+++ pyproject.toml
@@ -91,6 +91,12 @@ mike = "^2.1.2"
 
 
 
+
+[tool.poetry.group.opentelemetry.dependencies]
+opentelemetry-api = "^1.32.1"
+opentelemetry-sdk = "^1.32.1"
+opentelemetry-exporter-otlp = "^1.32.1"
+
 [pytest]
 env_files = [".env"]
 
