+++ ecologits/utils/opentelemetry.py
@@ -1,13 +1,12 @@
-import uuid
 from opentelemetry import metrics
 from opentelemetry.exporter.otlp.proto.http.metric_exporter import OTLPMetricExporter
 from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.metrics.export import PeriodicExportingMetricReader
 
 
-class EcologitsTelemetry:
-    def __init__(self, collector_url: str):
-        exporter = OTLPMetricExporter(endpoint=collector_url)
+class OpenTelemetry:
+    def __init__(self, endpoint: str):
+        exporter = OTLPMetricExporter(endpoint=endpoint)
         reader = PeriodicExportingMetricReader(exporter, export_interval_millis=5000)
         provider = MeterProvider(metric_readers=[reader])
         metrics.set_meter_provider(provider)
@@ -19,17 +18,17 @@ class EcologitsTelemetry:
             description="Total number of AI requests",
             unit="1"
         )
-        self.tokens_input = meter.create_counter(
-            name="ecologits_tokens_input_total",
+        self.input_tokens = meter.create_counter(
+            name="ecologits_input_tokens",
             description="Total input tokens",
             unit="1"
         )
-        self.tokens_output = meter.create_counter(
-            name="ecologits_tokens_output_total",
+        self.output_tokens = meter.create_counter(
+            name="ecologits_output_tokens",
             description="Total output tokens",
             unit="1"
         )
-        self.latency_hist = meter.create_counter(
+        self.request_latency = meter.create_counter(
             name="ecologits_request_latency_seconds",
             description="Request latency in seconds",
             unit="s"
@@ -55,22 +54,28 @@ class EcologitsTelemetry:
             unit="MJ"
         )
 
-    def record_request(self, prompt_tokens: int, completion_tokens: int, latency_s: float, 
-                       energy_value: float, gwp_value: float, adpe_value: float, 
-                       pe_value: float, model="gpt-4", endpoint="/ask") -> None:
-        #request_id = str(uuid.uuid4())
-
+    def record_request(
+            self,
+            input_tokens: int,
+            output_tokens: int,
+            request_latency: float,
+            energy_value: float,
+            gwp_value: float,
+            adpe_value: float,
+            pe_value: float,
+            model: str,
+            endpoint: str
+    ) -> None:
         labels = {
             "endpoint": endpoint,
             "model": model
         }
 
         self.request_counter.add(1, labels)
-        self.tokens_input.add(prompt_tokens, labels)
-        self.tokens_output.add(completion_tokens, labels)
-        self.latency_hist.add(float(latency_s), labels)
+        self.input_tokens.add(input_tokens, labels)
+        self.output_tokens.add(output_tokens, labels)
+        self.request_latency.add(float(request_latency), labels)
         self.energy_value.add(float(energy_value.max), labels)
         self.gwp_value.add(float(gwp_value.max), labels)
         self.adpe_value.add(float(adpe_value.max), labels)
         self.pe_value.add(float(pe_value.max), labels)
-            
