+++ paper/paper.md
@@ -44,7 +44,7 @@ In parallel to the large deployment of energy-consuming models in the industry, 
 
 # Methodology
 
-The **Attributional Life Cycle Assessment (A-LCA)** methodology defined by the ISO 14044 standard [@klopffer_life_1997; @boustead_lca_1996; @hunt_lca_1996] is used to properly estimate potential environmental impacts of products, processes, projects, or services. In EcoLogits, we apply this methodology to estimate the impacts of a request to a GenAI inference service. The methodology encompasses a multi-phase and multi-criteria environmental impacts.
+The **Attributional Life Cycle Assessment (A-LCA)** methodology defined by the ISO 14044 standard [@klopffer_life_1997; @boustead_lca_1996; @hunt_lca_1996] is used to properly estimate potential environmental impacts of products, processes, projects, or services. In EcoLogits, we apply this methodology to estimate the impacts of a request to a GenAI inference service. The methodology encompasses multi-phase and multi-criteria environmental impacts.
 
 In our approach we consider the usage phase accounting for the direct energy consumption [@llm-perf-leaderboard; @optimum-benchmark] and the embodied phase accounting for the hardware production, including raw material extraction, manufacturing, and transportation [@boaviztapi]. We do not include the end-of-life phase for lack of data and transparency on e-waste collection and recycling [@bordage_digital_2021; @forti_global_2020]. EcoLogits reports an estimation of the direct energy consumption in addition to other environmental impact criteria such as *Global Warming Potential* [@ipcc_2013], *Abiotic Depletion Potential* [@van_oers_abiotic_2020] and *Primary Energy Consumption* [@frischknecht_cumulative_2015]. Assessing multiple impact criteria helps prevent pollution shifting that can result from optimizing for a single criterion.
 
