### 2025-05-13 #7471 @Zeitsperre
@Fei-Tao Hi there, excuse the late response!

I've completed my review of `ecologits` and opened a few issues to address the problems that I felt weren't covered by @jaywonchung's review (https://github.com/genai-impact/ecologits/issues/95).

Once the issues linked above this comment are addressed, consider my change requests satisfied.

@adrienbanse C'est un outil très fascinant et je vois maintenant les défis liés à l'estimation des impacts environnementaux des systèmes opaques. J'espère que les changements à faire ici nécessitent pas trop d'efforts de votre part et de votre équipe. Merci encore pour votre contribution à JOSS!
