### 2025-05-19 a9a3b69
fix: change min/max value selection in llm dab
