### 2025-05-19 cd8b6f7
chore: bump version to 0.7.0
