+++ ecologits/__init__.py
@@ -1,6 +1,6 @@
 from ._ecologits import EcoLogits
 
-__version__ = "0.6.3"
+__version__ = "0.7.0"
 __all__ = [
     "EcoLogits",
     "__version__"
