### 2025-05-28 7dbc9b8
docs: switch to llama example
