### 2025-05-28 c67262e
docs: add example and correct some mistakes
