+++ docs/index.md
@@ -61,44 +61,120 @@ For detailed instructions on each provider, refer to the complete list of [suppo
 
 ## Usage Example
 
-Below is a simple example demonstrating how to use the GPT-4o-mini model from OpenAI with **EcoLogits** to track environmental impacts.
+Below are simple examples demonstrating how to use LLM models with **EcoLogits** to track environmental impacts. Go to the [tutorial page](tutorial/providers.md) for more complete examples of each supported provider. 
 
-```python
-from ecologits import EcoLogits
-from openai import OpenAI
+=== "OpenAI"
 
-# Initialize EcoLogits
-EcoLogits.init()
+    ```python
+    from ecologits import EcoLogits
+    from openai import OpenAI
 
-client = OpenAI(api_key="<OPENAI_API_KEY>")
+    # Initialize EcoLogits
+    EcoLogits.init()
 
-response = client.chat.completions.create( # (1)!
-    model="gpt-4o-mini",
-    messages=[
-        {"role": "user", "content": "Tell me a funny joke!"}
-    ]
-)
+    client = OpenAI(api_key="<OPENAI_API_KEY>")
 
-# Get estimated environmental impacts of the inference
-print(f"Energy consumption: {response.impacts.energy.value} kWh") # (2)!
-print(f"GHG emissions: {response.impacts.gwp.value} kgCO2eq") # (3)!
+    response = client.chat.completions.create( # (1)!
+        model="gpt-4o-mini",
+        messages=[
+            {"role": "user", "content": "Tell me a funny joke!"}
+        ]
+    )
 
-# Get potential warnings
-if response.impacts.has_warnings:
-    for w in response.impacts.warnings:
-        print(w) # (4)!
+    # Get estimated environmental impacts of the inference
+    print(f"Energy consumption: {response.impacts.energy.value} kWh") # (2)!
+    print(f"GHG emissions: {response.impacts.gwp.value} kgCO2eq") # (3)!
 
-# Get potential errors
-if response.impacts.has_errors:
-    for w in response.impacts.errors:
-        print(w) # (5)!
-```
+    # Get potential warnings
+    if response.impacts.has_warnings:
+        for w in response.impacts.warnings:
+            print(w) # (4)!
 
-1. You don't need to change your code when making a request! :tada:
-2. Total estimated energy consumption for the request in kilowatt-hour (kWh). You can expect an interval, see [exemple here](tutorial/impacts.md#example-with-a-rangevalue). 
-3. Total estimated greenhouse gas emissions for the request in kilogram of CO2 equivalent (kgCO2eq). You can expect an interval, see [exemple here](tutorial/impacts.md#example-with-a-rangevalue). 
-4. For `gpt-4o-mini`, you can expect two warnings: [`model-arch-not-released`](tutorial/warnings_and_errors.md#model-arch-not-released) and [`model-arch-multimodal`](tutorial/warnings_and_errors.md#model-arch-multimodal).
-5. On this example you shouldn't get any errors.
+    # Get potential errors
+    if response.impacts.has_errors:
+        for w in response.impacts.errors:
+            print(w) # (5)!
+    ```
+
+    1. You don't need to change your code when making a request! :tada:
+    2. Total estimated energy consumption for the request in kilowatt-hour (kWh). You can expect an interval, see [example here](tutorial/impacts.md#example-with-a-rangevalue). 
+    3. Total estimated greenhouse gas emissions for the request in kilogram of CO2 equivalent (kgCO2eq). You can expect an interval, see [example here](tutorial/impacts.md#example-with-a-rangevalue). 
[…]
