### 2025-05-28 4fa2c4e
doc: add more API descrptions
