### 2025-05-28 bb8e7b0
doc: add more API descrptions
