+++ ecologits/tracers/mistralai_tracer_v0.py
@@ -19,10 +19,16 @@ PROVIDER = "mistralai"
 
 
 class ChatCompletionResponse(_ChatCompletionResponse):
+    """
+        Wrapper of `mistralai.models.chat_completion.ChatCompletionResponse` with `ImpactsOutput`
+    """
     impacts: ImpactsOutput
 
 
 class ChatCompletionStreamResponse(_ChatCompletionStreamResponse):
+    """
+        Wrapper of `mistralai.models.chat_completion.ChatCompletionStreamResponse` with `ImpactsOutput`
+    """
     impacts: ImpactsOutput
 
 
@@ -30,6 +36,19 @@ class ChatCompletionStreamResponse(_ChatCompletionStreamResponse):
 def mistralai_chat_wrapper(
     wrapped: Callable, instance: MistralClient, args: Any, kwargs: Any  # noqa: ARG001
 ) -> ChatCompletionResponse:
+    """
+    Function that wraps a MistralAI answer with computed impacts 
+
+    Args:
+        wrapped: Callable that returns the LLM response
+        instance: Never used - for compatibility with `wrapt`
+        args: Arguments of the callable
+        kwargs: Keyword arguments of the callable
+
+    Returns:
+        A wrapped `ChatCompletionResponse` with impacts
+    """
+
     timer_start = time.perf_counter()
     response = wrapped(*args, **kwargs)
     request_latency = time.perf_counter() - timer_start
@@ -49,6 +68,19 @@ def mistralai_chat_wrapper(
 def mistralai_chat_wrapper_stream(
     wrapped: Callable, instance: MistralClient, args: Any, kwargs: Any  # noqa: ARG001
 ) -> Iterable[ChatCompletionStreamResponse]:
+    """
+    Function that wraps a MistralAI answer with computed impacts in streaming mode
+
+    Args:
+        wrapped: Callable that returns the LLM response
+        instance: Never used - for compatibility with `wrapt`
+        args: Arguments of the callable
+        kwargs: Keyword arguments of the callable
+
+    Returns:
+        A wrapped `Iterable[ChatCompletionStreamResponse]` with impacts
+    """
+
     timer_start = time.perf_counter()
     stream = wrapped(*args, **kwargs)
     token_count = 0
@@ -76,6 +108,18 @@ async def mistralai_async_chat_wrapper(
     args: Any,
     kwargs: Any,
 ) -> ChatCompletionResponse:
+    """
+    Function that wraps a MistralAI answer with computed impacts in async mode
+
+    Args:
+        wrapped: Async callable that returns the LLM response
+        instance: Never used - for compatibility with `wrapt`
+        args: Arguments of the callable
+        kwargs: Keyword arguments of the callable
+
+    Returns:
+        A wrapped `ChatCompletionResponse` with impacts
+    """
     timer_start = time.perf_counter()
     response = await wrapped(*args, **kwargs)
     request_latency = time.perf_counter() - timer_start
@@ -98,6 +142,18 @@ async def mistralai_async_chat_wrapper_stream(
     args: Any,
     kwargs: Any,
 ) -> AsyncGenerator[ChatCompletionStreamResponse, None]:
+    """
+    Function that wraps a MistralAI answer with computed impacts in streaming and async mode
+
+    Args:
+        wrapped: Callable that returns the LLM response
+        instance: Never used - for compatibility with `wrapt`
+        args: Arguments of the callable
+        kwargs: Keyword arguments of the callable
+
+    Returns:
+        A wrapped `AsyncGenerator[ChatCompletionStreamResponse, None]` with impacts
+    """
     timer_start = time.perf_counter()
     stream = wrapped(*args, **kwargs)
     token_count = 0
[…]
