+++ ecologits/tracers/openai_tracer.py
@@ -14,10 +14,16 @@ PROVIDER = "openai"
 
 
 class ChatCompletion(_ChatCompletion):
+    """
+        Wrapper of `openai.types.chat.ChatCompletion` with `ImpactsOutput`
+    """
     impacts: ImpactsOutput
 
 
 class ChatCompletionChunk(_ChatCompletionChunk):
+    """
+        Wrapper of `openai.types.chat.ChatCompletionChunk` with `ImpactsOutput`
+    """
     impacts: ImpactsOutput
 
 
@@ -27,6 +33,18 @@ def openai_chat_wrapper(
     args: Any,
     kwargs: Any
 ) -> Union[ChatCompletion, Stream[ChatCompletionChunk]]:
+    """
+    Function that wraps an OpenAI answer with computed impacts 
+
+    Args:
+        wrapped: Callable that returns the LLM response
+        instance: Never used - for compatibility with `wrapt`
+        args: Arguments of the callable
+        kwargs: Keyword arguments of the callable
+
+    Returns:
+        A wrapped `ChatCompletion` or `Stream[ChatCompletionChunk]` with impacts
+    """
     if kwargs.get("stream", False):
         return openai_chat_wrapper_stream(wrapped, instance, args, kwargs)
     else:
@@ -92,6 +110,18 @@ async def openai_async_chat_wrapper(
     args: Any,
     kwargs: Any,
 ) -> Union[ChatCompletion, AsyncStream[ChatCompletionChunk]]:
+    """
+    Function that wraps an OpenAI answer with computed impacts in async mode
+
+    Args:
+        wrapped: Async callable that returns the LLM response
+        instance: Never used - for compatibility with `wrapt`
+        args: Arguments of the callable
+        kwargs: Keyword arguments of the callable
+
+    Returns:
+        A wrapped `ChatCompletion` or `AsyncStream[ChatCompletionChunk]` with impacts
+    """
     if kwargs.get("stream", False):
         return openai_async_chat_wrapper_stream(wrapped, instance, args, kwargs)
     else:
@@ -153,6 +183,9 @@ async def openai_async_chat_wrapper_stream(  # type: ignore[misc]
 
 
 class OpenAIInstrumentor:
+    """
+        Instrumentor initialized by EcoLogits to automatically wrap all OpenAI calls
+    """
     def __init__(self) -> None:
         self.wrapped_methods = [
             {
