+++ ecologits/tracers/mistralai_tracer_v0.py
@@ -37,7 +37,7 @@ def mistralai_chat_wrapper(
     wrapped: Callable, instance: MistralClient, args: Any, kwargs: Any  # noqa: ARG001
 ) -> ChatCompletionResponse:
     """
-    Function that wraps a MistralAI answer with computed impacts 
+    Function that wraps a MistralAI answer with computed impacts
 
     Args:
         wrapped: Callable that returns the LLM response
