+++ ecologits/tracers/mistralai_tracer_v1.py
@@ -33,7 +33,7 @@ def mistralai_chat_wrapper(
     wrapped: Callable, instance: Mistral, args: Any, kwargs: Any  # noqa: ARG001
 ) -> ChatCompletionResponse:
     """
-    Function that wraps a MistralAI answer with computed impacts 
+    Function that wraps a MistralAI answer with computed impacts
 
     Args:
         wrapped: Callable that returns the LLM response
