### 2025-06-02 7fa4323
docs: add a new sponsor
