### 2025-06-02 14d2e84
docs: minor corrections
