### 2025-06-02 368de3b
docs: minor corrections
