### 2025-06-04 d43b296
fix when int given and not float
