### 2025-06-05 492c5e6
doc: update readme with new DOI
