### 2025-06-05 e8101f6
remove samuel affiliation
