### 2025-06-14 d9068e1
docs: add a new sponsor
