### 2025-06-14 f72fbab
fix: use mean instead of avg for range values
Recently merged a PR that implements mean property on RangeValue. We can remove our own implementation of avg.
