### 2025-06-14 a54c709
test: add opentelementry tests
