+++ tests/test_opentelemetry.py
@@ -1,0 +1,186 @@
+import pytest
+from unittest.mock import Mock, patch, MagicMock
+from ecologits.impacts.modeling import GWP, ADPe, PE, Energy
+from ecologits.utils.range_value import RangeValue
+from ecologits.utils.opentelemetry import OpenTelemetry
+
+
+@pytest.fixture
+def mock_opentelemetry_setup():
+    """Fixture to mock all OpenTelemetry dependencies"""
+    with patch('ecologits.utils.opentelemetry.OTLPMetricExporter') as mock_exporter_class, \
+            patch('ecologits.utils.opentelemetry.PeriodicExportingMetricReader') as mock_reader_class, \
+            patch('ecologits.utils.opentelemetry.MeterProvider') as mock_provider_class, \
+            patch('ecologits.utils.opentelemetry.metrics') as mock_metrics:
+        # Create mock instances
+        mock_exporter = Mock()
+        mock_reader = Mock()
+        mock_provider = Mock()
+        mock_meter = Mock()
+
+        # Configure the mocks
+        mock_exporter_class.return_value = mock_exporter
+        mock_reader_class.return_value = mock_reader
+        mock_provider_class.return_value = mock_provider
+        mock_metrics.get_meter.return_value = mock_meter
+
+        # Mock counter creation
+        mock_counter = Mock()
+        mock_meter.create_counter.return_value = mock_counter
+
+        yield {
+            'exporter_class': mock_exporter_class,
+            'reader_class': mock_reader_class,
+            'provider_class': mock_provider_class,
+            'metrics': mock_metrics,
+            'meter': mock_meter,
+            'counter': mock_counter
+        }
+
+
+def test_init(mock_opentelemetry_setup):
+    """Test the initialization of OpenTelemetry class"""
+    endpoint = "http://localhost:4318/v1/metrics"
+    telemetry = OpenTelemetry(endpoint)
+
+    # Verify OTLPMetricExporter was initialized with the correct endpoint
+    mock_opentelemetry_setup['exporter_class'].assert_called_once_with(endpoint=endpoint)
+
+    # Verify PeriodicExportingMetricReader was initialized correctly
+    mock_opentelemetry_setup['reader_class'].assert_called_once()
+
+    # Verify MeterProvider was initialized with the reader
+    mock_opentelemetry_setup['provider_class'].assert_called_once()
+
+    # Verify metrics provider was set
+    mock_opentelemetry_setup['metrics'].set_meter_provider.assert_called_once()
+
+    # Verify meter was created
+    mock_opentelemetry_setup['metrics'].get_meter.assert_called_once_with("ecologits")
+
+    # Verify all counters were created (8 counters in total)
+    assert mock_opentelemetry_setup['meter'].create_counter.call_count == 8
+
+
+def test_record_request_with_valid_data(mock_opentelemetry_setup):
+    """Test record_request method with valid data"""
+    telemetry = OpenTelemetry("http://localhost:4318/v1/metrics")
+
+    # Reset the mock to clear the initialization calls
+    mock_counter = mock_opentelemetry_setup['counter']
+    mock_counter.add.reset_mock()
+
+    # Create test data
+    input_tokens = 100
+    output_tokens = 50
+    request_latency = 1.5
+    energy = Energy(value=0.1)
+    gwp = GWP(value=0.2)
+    adpe = ADPe(value=0.3)
+    pe = PE(value=0.4)
+    model = "gpt-4o-mini"
+    endpoint = "openai"
+
+    # Call the method
+    telemetry.record_request(
+        input_tokens=input_tokens,
[…]
