### 2025-06-14 56cfc69
test: add opentelementry tests
