### 2025-06-16 1239ceb
feat: add deprecation warning when initializing without providers defined.
