+++ README.md
@@ -34,7 +34,7 @@ from ecologits import EcoLogits
 from openai import OpenAI
 
 # Initialize EcoLogits
-EcoLogits.init()
+EcoLogits.init(providers=["openai"])
 
 client = OpenAI(api_key="<OPENAI_API_KEY>")
 
@@ -46,8 +46,8 @@ response = client.chat.completions.create(
 )
 
 # Get estimated environmental impacts of the inference
-print(f"Energy consumption: {response.impacts.energy.value} kWh")
-print(f"GHG emissions: {response.impacts.gwp.value} kgCO2eq")
+print(f"Energy consumption: {response.impacts.energy.value.mean} kWh")
+print(f"GHG emissions: {response.impacts.gwp.value.mean} kgCO2eq")
 ```
 
 See package documentation on [EcoLogits](https://ecologits.ai/)
