### 2025-06-16 407a79b
feat: add deprecation warning when initializing without providers defined.
