### 2025-06-16 5ef488d
feat: add deprecation warning when initializing without providers defined.
