### 2025-06-16 978ed1e
docs: update init in all docs
