### 2025-06-17 45264c7
feat: use impacts instead of individual impact values
