+++ ecologits/utils/opentelemetry.py
@@ -1,12 +1,10 @@
-from typing import Optional
-
 from opentelemetry import metrics
 from opentelemetry.exporter.otlp.proto.http.metric_exporter import OTLPMetricExporter
 from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.metrics.export import PeriodicExportingMetricReader
 
-from ecologits.impacts.modeling import GWP, PE, ADPe, Energy
 from ecologits.log import logger
+from ecologits.tracers.utils import ImpactsOutput
 from ecologits.utils.range_value import RangeValue
 
 
@@ -65,17 +63,14 @@ class OpenTelemetry:
             input_tokens: int,
             output_tokens: int,
             request_latency: float,
-            energy: Optional[Energy],
-            gwp: Optional[GWP],
-            adpe: Optional[ADPe],
-            pe: Optional[PE],
+            impacts: ImpactsOutput,
             model: str,
             endpoint: str
     ) -> None:
-        if energy is None \
-                or gwp is None \
-                or adpe is None \
-                or pe is None:
+        if impacts.energy is None \
+                or impacts.gwp is None \
+                or impacts.adpe is None \
+                or impacts.pe is None:
             logger.error("Skipped sending request metrics because one of the impact values is None.")
             return
 
@@ -84,10 +79,11 @@ class OpenTelemetry:
             "model": model
         }
 
-        energy_value = energy.value.mean if isinstance(energy.value, RangeValue) else energy.value
-        gwp_value = gwp.value.mean if isinstance(gwp.value, RangeValue) else gwp.value
-        adpe_value = adpe.value.mean if isinstance(adpe.value, RangeValue) else adpe.value
-        pe_value = pe.value.mean if isinstance(pe.value, RangeValue) else pe.value
+        energy_value = impacts.energy.value.mean if isinstance(impacts.energy.value,
+                                                               RangeValue) else impacts.energy.value
+        gwp_value = impacts.gwp.value.mean if isinstance(impacts.gwp.value, RangeValue) else impacts.gwp.value
+        adpe_value = impacts.adpe.value.mean if isinstance(impacts.adpe.value, RangeValue) else impacts.adpe.value
+        pe_value = impacts.pe.value.mean if isinstance(impacts.pe.value, RangeValue) else impacts.pe.value
 
         self.request_counter.add(1, labels)
         self.input_tokens.add(input_tokens, labels)
