### 2025-06-17 21a37b8
feat: use impacts instead of individual impact values
