### 2025-06-17 281cb83
feat: use impacts instead of individual impact values
