### 2025-06-17 75e06ca
feat: change error message
