### 2025-06-17 2ef331a
feat: add support for stream, async and async+stream monitoring with openai
