### 2025-06-17 1f0fbaa
docs: fix a typo in async stream for openai
