### 2025-06-19 8b36706
feat: allow for empty repo
