### 2025-06-19 1346030
tests: write tests for empty + custom models
