+++ tests/test_model_repository.py
@@ -1,5 +1,10 @@
 from ecologits.model_repository import ModelRepository
 
+def test_create_empty_repository():
+    models = ModelRepository()
+    assert isinstance(models, ModelRepository)
+    assert not models.list_models()
+
 
 def test_create_model_repository_default():
     models = ModelRepository.from_json()
@@ -23,3 +28,17 @@ def test_ambiguous_names():
     assert models.find_model(provider="openai", model_name="gpt-4-turbo").name == "gpt-4-turbo"
     assert models.find_model(provider="openai", model_name="gpt-4o").name == "gpt-4o"
     assert models.find_model(provider="openai", model_name="gpt-35-turbo").name == "gpt-35-turbo"
+
+
+def test_add_custom_model():
+    models = ModelRepository()
+    custom_model_data = {
+        "provider": "openai",
+        "name": "gpt-4.1-2025-04-14",
+        "architecture": {
+            "type": "moe", 
+            "parameters": 1000
+        }
+    }
+    models.add_model(custom_model_data)
+    assert models.find_model("openai", "gpt-4.1-2025-04-14") is not None    
