### 2025-06-25 e991c2b
fix: use a deprecation warning instead
