### 2025-06-30 a87f994
fix: missing DOI
