+++ paper/paper.bib
@@ -212,18 +212,23 @@
   title = {Optimum-Benchmark: A framework for benchmarking the performance of Transformers models with different hardwares, backends and optimizations.},
 }
 
-@inproceedings{boaviztapi,
-  title = {{BoaviztAPI: a bottom-up model to assess the environmental impacts of cloud services}},
-  author = {Simon, Thibault and Ekchajzer, David and Berthelot, Adrien and Fourboul, Eric and Rince, Samuel and Rouvoy, Romain},
-  url = {https://hal.science/hal-04621947},
-  booktitle = {{HotCarbon'24 - 3rd Workshop on Sustainable Computer Systems}},
-  address = {Santa Cruz, United States},
-  year = {2024},
-  month = Jul,
-  keywords = {Life Cycle Assessment ; Bottom-up ; Multicriteria ; Open-data ; Cloud ; Environmental Impact},
-  pdf = {https://hal.science/hal-04621947v2/file/HotCarbon_hal.pdf},
-  hal_id = {hal-04621947},
-  hal_version = {v2},
+@article{boaviztapi,
+author = {Simon, Thibault and Ekchajzer, David and Berthelot, Adrien and Fourboul, Eric and Rince, Samuel and Rouvoy, Romain},
+title = {BoaviztAPI: A Bottom-Up Model to Assess the Environmental Impacts of Cloud Services},
+year = {2025},
+issue_date = {December 2024},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {4},
+number = {5},
+url = {https://doi.org/10.1145/3727200.3727213},
+doi = {10.1145/3727200.3727213},
+abstract = {Awareness surrounding the environmental impacts of the digital industry has led numerous professionals to incorporate these considerations into their work. However, the conceptualization of environmental impacts has often been narrowed to the scope of carbon footprint. This limitation can be attributed to various technical and data accessibility constraints, hindering a comprehensive evaluation, including a multi-criteria analysis over the entire life cycle of digital technologies.In response to these limitations, we introduce a comprehensive bottom-up evaluation method suitable for servers and cloud instances, employing a life cycle thinking approach. We start by modeling the life cycle impacts of a server based on its hardware configuration. Then, we aggregate these with the impacts of its technical and physical environment to define the impacts of a cloud platform. We finally model the impacts of a cloud instance as a portion of the cloud platform.The proposed approach has been implemented as an open-source toolkit and published as an API. This initiative aims to provide developers and researchers with a tool for conducting environmental evaluations of their infrastructure based on open data and open methodologies, enhancing their ability to explore the environmental materiality of ICT products, services, and infrastructures.},
+journal = {SIGENERGY Energy Inform. Rev.},
+month = apr,
+pages = {84–90},
+numpages = {7},
+keywords = {cloud, environmental impact, life cycle assessment, bottom-up, multicriteria, open-data}
 }
 
 @article{klopffer_life_1997,
