### 2025-07-08 9fa9696
Update README.md
update sponsors to sponsors & benefactors
