+++ README.md
@@ -51,8 +51,7 @@ print(f"GHG emissions: {response.impacts.gwp.value} kgCO2eq")
 
 See package documentation on [EcoLogits](https://ecologits.ai/)
 
-## 💚 Sponsors
-
+## 💚 Sponsors & benefactors
 
 <a href="https://resilio-solutions.com/" target="_blank">
 <img src="https://raw.githubusercontent.com/genai-impact/ecologits/main/docs/assets/sponsors/resilio.png" alt="Resilio" height="100" width="250">
@@ -66,6 +65,9 @@ See package documentation on [EcoLogits](https://ecologits.ai/)
 <img src="https://raw.githubusercontent.com/genai-impact/ecologits/main/docs/assets/sponsors/sopht.png" alt="Sopht" height="100" width="250">
 </a>
 
+<a href="https://www.culture.gouv.fr/fr/" target="_blank">
+<img src="https://raw.githubusercontent.com/genai-impact/ecologits/main/docs/assets/sponsors/ministereculture.png" alt="Ministere de la culture" height="100" width="250">
+</a>
 
 ## 💪 Contributing
 
