### 2025-07-08 888cb58
fix: minor typo in RELEASE.md note
