### 2025-07-08 3f70f54
chore: bump version to 0.7.2
