+++ ecologits/__init__.py
@@ -1,6 +1,6 @@
 from ._ecologits import EcoLogits
 
-__version__ = "0.7.1"
+__version__ = "0.7.2"
 __all__ = [
     "EcoLogits",
     "__version__"
