### 2025-07-08 fb499fc
Update index.md
Update logo and text size
