### 2025-07-08 #7471 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
