### 2025-07-09 5848a4e
tests: write inmem otel tests
