+++ tests/test_opentelemetry_in_memory.py
@@ -1,0 +1,227 @@
+import pytest
+from unittest.mock import patch
+from opentelemetry.sdk.metrics import MeterProvider
+from opentelemetry.sdk.metrics.export import InMemoryMetricReader
+from opentelemetry import metrics
+from ecologits.impacts.modeling import GWP, ADPe, PE, Energy
+from ecologits.utils.range_value import RangeValue
+from ecologits.utils.opentelemetry import OpenTelemetry
+from ecologits.tracers.utils import ImpactsOutput
+
+
+@pytest.fixture
+def in_memory_telemetry():
+    """Create OpenTelemetry instance with in-memory metric reader for testing."""
+
+    # Create in-memory reader to capture metrics
+    reader = InMemoryMetricReader()
+    provider = MeterProvider(metric_readers=[reader])
+    
+    # Patch the OpenTelemetry class to use our in-memory setup
+    with patch('ecologits.utils.opentelemetry.OTLPMetricExporter'), \
+            patch('ecologits.utils.opentelemetry.PeriodicExportingMetricReader') as mock_reader, \
+            patch('ecologits.utils.opentelemetry.MeterProvider') as mock_provider, \
+            patch('ecologits.utils.opentelemetry.metrics') as mock_metrics:
+        
+        # Configure mocks to use our in-memory setup
+        mock_reader.return_value = reader
+        mock_provider.return_value = provider
+        mock_metrics.set_meter_provider.side_effect = lambda p: metrics.set_meter_provider(provider)
+        mock_metrics.get_meter.side_effect = lambda name: provider.get_meter(name)
+        
+        # Create telemetry instance
+        telemetry = OpenTelemetry("http://fake-endpoint:4318/v1/metrics")
+        
+        yield telemetry, reader
+
+
+def get_metric_data(reader, metric_name):
+    """Helper to extract metric data from the in-memory reader."""
+    metrics_data = reader.get_metrics_data()
+    if not metrics_data:
+        return None
+    
+    for resource_metrics in metrics_data.resource_metrics:
+        for scope_metrics in resource_metrics.scope_metrics:
+            for metric in scope_metrics.metrics:
+                if metric.name == metric_name:
+                    return metric
+    return None
+
+
+def get_metric_value_with_attributes(metric, expected_attributes):
+    """Helper to get metric value for specific attributes."""
+    if not metric or not metric.data or not metric.data.data_points:
+        return None
+    
+    for data_point in metric.data.data_points:
+        # Convert attributes to dict for comparison
+        point_attributes = dict(data_point.attributes) if data_point.attributes else {}
+        if point_attributes == expected_attributes:
+            return data_point.value
+    return None
+
+
+def test_init(in_memory_telemetry):
+    telemetry, reader = in_memory_telemetry
+    
+    # Just verify that the telemetry instance was created successfully
+    assert telemetry is not None
+    assert reader is not None
+    
+    # The counters should be accessible as attributes
+    assert hasattr(telemetry, 'request_counter')
+    assert hasattr(telemetry, 'input_tokens')
+    assert hasattr(telemetry, 'output_tokens')
+    assert hasattr(telemetry, 'request_latency')
+    assert hasattr(telemetry, 'energy_value')
+    assert hasattr(telemetry, 'gwp_value')
+    assert hasattr(telemetry, 'adpe_value')
+    assert hasattr(telemetry, 'pe_value')
+
+
+def test_record_request_with_valid_data(in_memory_telemetry):
+    telemetry, reader = in_memory_telemetry
+    
+    # Create test data
+    input_tokens = 100
[…]
