### 2025-07-09 cc6fdaa
feat: add contextmgr to label metrics
