### 2025-07-09 a38a1d9
tests: add otel label tests
