+++ tests/test_opentelemetry_in_memory.py
@@ -5,7 +5,7 @@ from opentelemetry.sdk.metrics.export import InMemoryMetricReader
 from opentelemetry import metrics
 from ecologits.impacts.modeling import GWP, ADPe, PE, Energy
 from ecologits.utils.range_value import RangeValue
-from ecologits.utils.opentelemetry import OpenTelemetry
+from ecologits.utils.opentelemetry import OpenTelemetry, otel_labels, get_current_labels
 from ecologits.tracers.utils import ImpactsOutput
 
 
@@ -225,3 +225,202 @@ def test_record_request_with_missing_data(in_memory_telemetry):
     # Verify still no counters were updated
     request_metric = get_metric_data(reader, "ecologits_requests_total")
     assert request_metric is None or len(request_metric.data.data_points) == 0
+
+def test_otel_labels_context_basic_functionality():
+    """Test that otel_labels context manager correctly stores and retrieves labels."""
+    # Test outside context - should be empty
+    assert get_current_labels() == {}
+    
+    # Test inside context
+    with otel_labels(user_id="user123", experiment="test"):
+        labels = get_current_labels()
+        assert labels == {"user_id": "user123", "experiment": "test"}
+    
+    # Test after context - should be empty again
+    assert get_current_labels() == {}
+
+
+def test_otel_labels_nested_contexts():
+    """Test nested otel_labels contexts merge correctly."""
+    with otel_labels(experiment="main", version="1.0"):
+        # First level
+        assert get_current_labels() == {"experiment": "main", "version": "1.0"}
+        
+        with otel_labels(user_id="user123", experiment="override"):
+            # Second level - experiment should be overridden
+            expected = {"experiment": "override", "version": "1.0", "user_id": "user123"}
+            assert get_current_labels() == expected
+            
+            with otel_labels(temp_flag=True):
+                # Third level - add more labels
+                expected = {
+                    "experiment": "override", 
+                    "version": "1.0", 
+                    "user_id": "user123",
+                    "temp_flag": True
+                }
+                assert get_current_labels() == expected
+            
+            # Back to second level
+            expected = {"experiment": "override", "version": "1.0", "user_id": "user123"}
+            assert get_current_labels() == expected
+        
+        # Back to first level
+        assert get_current_labels() == {"experiment": "main", "version": "1.0"}
+
+
+def test_record_request_with_user_labels(in_memory_telemetry):
+    """Test that user labels from context are included in metrics."""
+    telemetry, reader = in_memory_telemetry
+    
+    # Record request with user labels
+    with otel_labels(user_id="user123", experiment="test_run"):
+        telemetry.record_request(
+            input_tokens=200,
+            output_tokens=100,
+            request_latency=2.0,
+            impacts=ImpactsOutput(
+                energy=Energy(value=0.2),
+                gwp=GWP(value=0.4),
+                adpe=ADPe(value=0.6),
+                pe=PE(value=0.8)
+            ),
+            model="gpt-4",
+            endpoint="openai"
+        )
+    
+    # Force collection
+    reader.collect()
+    
+    # Check that metrics include both system and user labels
+    expected_attributes = {
+        "endpoint": "openai", 
+        "model": "gpt-4",
+        "user_id": "user123",
+        "experiment": "test_run"
+    }
+    
+    request_metric = get_metric_data(reader, "ecologits_requests_total")
+    assert request_metric is not None
+    assert get_metric_value_with_attributes(request_metric, expected_attributes) == 1
+    
+    energy_metric = get_metric_data(reader, "ecologits_energy_total")
+    assert energy_metric is not None
[…]
