+++ paper/paper.bib
@@ -312,7 +312,7 @@ keywords = {cloud, environmental impact, life cycle assessment, bottom-up, multi
 }
 
 @article{boustead_lca_1996,
-  title        = {{LCA} — How it came about: — The beginning in the U.K.},
+  title        = {{LCA} — How it came about: — The beginning in the {U.K.}},
   volume       = {1},
   issn         = {1614-7502},
   OPTurl          = {https://doi.org/10.1007/BF02978943},
