### 2025-07-09 50685fa
:lipstick: Update index.md
change logo size
